% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_regression.R
\name{absolute.order}
\alias{absolute.order}
\title{Ordering of Predictors from Coefficient Estimates}
\usage{
absolute.order(beta)
}
\arguments{
\item{beta}{A vector of estimated regression coefficients.}
}
\value{
An ordering of the predictors.
}
\description{
This function orders the predictors by decreasing
  order of the magnitude of the estimated regression coefficient.
}
\examples{
### generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

### glmnet fit
library(glmnet)
beta.lasso = coef(cv.glmnet(X, y))[-1]
lasso.order = absolute.order(beta.lasso)

### ncvreg fit
library(ncvreg)
beta.scad = c(coef(cv.ncvreg(X, y))[-1])
scad.order = absolute.order(beta.scad)

}
