% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss.R
\name{check_projection}
\alias{check_projection}
\title{Check whether b in space spanned by the non-zero eigenvectors of A}
\usage{
check_projection(A, b)
}
\arguments{
\item{A}{a p by p matrix}

\item{b}{a length p vector}
}
\value{
a list of result: \cr
\item{status}{whether b in space spanned by the non-zero eigenvectors of A}
\item{msg}{msg gives the difference between the projected b and b if status is FALSE}
}
\description{
Check whether b in space spanned by the non-zero eigenvectors of A
}
