% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss.R
\name{check_r_matrix}
\alias{check_r_matrix}
\title{Check input covariance / correlation matrix}
\usage{
check_r_matrix(R, expected_dim, r_tol)
}
\arguments{
\item{R}{a p by p matrix of X'X, covariance matrix or correlation matrix}
}
\value{
a verified matrix
}
\description{
Check input covariance / correlation matrix
}
\keyword{internal}
