% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.ash.R
\name{coef.mr.ash}
\alias{coef.mr.ash}
\title{Extract Regression Coefficients from Mr.ASH Fit}
\usage{
\method{coef}{mr.ash}(object, ...)
}
\arguments{
\item{object}{A Mr.ASH fit, usually the result of calling
\code{mr.ash}.}

\item{...}{Additional arguments passed to the default S3 method.}
}
\value{
A p+1 vector. The first element gives the estimated
  intercept, and the remaining p elements are the estimated
  regression coefficients.
  
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X %*% beta + rnorm(n)

## fit mr.ash
fit.mr.ash  = mr.ash(X, y)

## coefficient
coef.mr.ash = coef(fit.mr.ash)
intercept   = coef.mr.ash[1]
beta        = coef.mr.ash[-1]
}
\description{
Retrieve posterior mean estimates of the regression
  coefficients in a Mr.ASH model.
}
