% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.susie.R
\name{coef.susie}
\alias{coef.susie}
\title{extract regression coefficients from susie fit}
\usage{
\method{coef}{susie}(s)
}
\arguments{
\item{s}{a susie fit}
}
\value{
a p vector of estimated regression coefficients
}
\description{
extract regression coefficients from susie fit
}
