% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_X_attributes.R
\name{compute_colSds}
\alias{compute_colSds}
\title{computes column standard deviations for any type of matrix
       replace matrixStats::colSds since this function only takes a dense matrix}
\usage{
compute_colSds(X)
}
\arguments{
\item{X}{an n by p matrix of any type, e.g. sparse, dense}
}
\value{
a p vector of column standard deviations
}
\description{
computes column standard deviations for any type of matrix
       replace matrixStats::colSds since this function only takes a dense matrix
}
