% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_Dinvb}
\alias{compute_Dinvb}
\title{Compute unscaled D-inverse \%*\% b using special structure of D-inverse}
\usage{
compute_Dinvb(order, b)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{b}{an n=p vector}
}
\value{
an n vector
}
\description{
Compute unscaled D-inverse \%*\% b using special structure of D-inverse
}
\keyword{internal}
