% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_Dinvty}
\alias{compute_Dinvty}
\title{Compute unscaled t(D-inverse) \%*\% y using special structure of D-inverse}
\usage{
compute_Dinvty(order, y)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{y}{an n vector}
}
\value{
an n vector
}
\description{
Compute unscaled t(D-inverse) \%*\% y using special structure of D-inverse
}
\keyword{internal}
