% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_multiplication.R
\name{compute_MXt}
\alias{compute_MXt}
\title{Compute M\%*\%t(scaled.X) using sparse multiplication}
\usage{
compute_MXt(M, X)
}
\arguments{
\item{M}{a L by p matrix}

\item{X}{is a scaled dense matrix or an unscaled sparse matrix}
}
\value{
a L by n matrix
}
\description{
Compute M\%*\%t(scaled.X) using sparse multiplication
}
