% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ss.R
\name{compute_ss}
\alias{compute_ss}
\title{Compute sufficient statistics from input data}
\usage{
compute_ss(X, y, standardize = TRUE)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{y}{an n vector}

\item{standardize}{logical flag (default=TRUE) for whether to standardize columns of X to unit variance prior to fitting.}
}
\value{
a list of sufficient statistics
}
\description{
Compute sufficient statistics from input data
}
