% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stumps_multiplication.R
\name{compute_stumps_Xb}
\alias{compute_stumps_Xb}
\title{Compute unscaled Xexpand \%*\% b where Xexpand is the n by np matrix formed by, for each of the p columns of Xrank form the n different stumps (changepoint basis functions), and concatentate}
\usage{
compute_stumps_Xb(Xord, b)
}
\arguments{
\item{Xord}{is an n by p matrix whose jth column containing order(X[,j])}

\item{b}{an np vector whose first n elements correspond to the first column of Xord, next n correspond to second column, etc.}
}
\value{
an n vector
}
\description{
Compute unscaled Xexpand \%*\% b where Xexpand is the n by np matrix formed by, for each of the p columns of Xrank form the n different stumps (changepoint basis functions), and concatentate
}
\keyword{internal}
