% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stumps_multiplication.R
\name{compute_stumps_Xty}
\alias{compute_stumps_Xty}
\title{Compute unscaled t(Xexpand) \%*\% y here Xexpand is the n by np matrix formed by, for each of the p columns of Xrank form the n different stumps (changepoint basis functions), and concatentate}
\usage{
compute_stumps_Xty(Xord, y)
}
\arguments{
\item{Xord}{is an n by p matrix whose jth column containing order(X[,j])}

\item{y}{an n vector}
}
\value{
an np vector
}
\description{
Compute unscaled t(Xexpand) \%*\% y here Xexpand is the n by np matrix formed by, for each of the p columns of Xrank form the n different stumps (changepoint basis functions), and concatentate
}
\keyword{internal}
