% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_tf_cm}
\alias{compute_tf_cm}
\title{Compute column mean of the trend filtering matrix X}
\usage{
compute_tf_cm(order, n)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{n}{the length of y}
}
\value{
an n vector
}
\description{
Compute column mean of the trend filtering matrix X
}
\keyword{internal}
