% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_tf_d}
\alias{compute_tf_d}
\title{Compute colSums(X * X) for X under four scenarios}
\usage{
compute_tf_d(order, n, cm, csd, standardize = FALSE, intercept = FALSE)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{n}{the length of y}

\item{cm}{column means of X}

\item{csd}{column standard deviations of X}

\item{standardize}{a boolean denotes whether scaling X by standard deviation}

\item{intercept}{a boolean denotes whether mean centering X}
}
\value{
an n vector
}
\description{
Compute colSums(X * X) for X under four scenarios
}
\keyword{internal}
