% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_tf_std_d}
\alias{compute_tf_std_d}
\title{A fast way to compute colSums(X * X), where X is a mean-centered and standardized trend filtering matrix}
\usage{
compute_tf_std_d(order, n)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{n}{the length of y}
}
\value{
an n vector
}
\description{
A fast way to compute colSums(X * X), where X is a mean-centered and standardized trend filtering matrix
}
\keyword{internal}
