% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_tf_X2tEb2}
\alias{compute_tf_X2tEb2}
\title{Compute sum(t(X * X) * Eb2), where X is a standardized trend filtering matrix}
\usage{
compute_tf_X2tEb2(X, Eb2)
}
\arguments{
\item{X}{a trend filtering matrix}

\item{Eb2}{an n vector}
}
\value{
a scalar
}
\description{
Compute sum(t(X * X) * Eb2), where X is a standardized trend filtering matrix
}
\keyword{internal}
