% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_tf_Xb}
\alias{compute_tf_Xb}
\title{Compute unscaled X \%*\% b using the special structure of trend filtering}
\usage{
compute_tf_Xb(order, b)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{b}{an n=p vector}
}
\value{
an n vector
}
\description{
Compute unscaled X \%*\% b using the special structure of trend filtering
}
\keyword{internal}
