% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{compute_tf_Xty}
\alias{compute_tf_Xty}
\title{Compute unscaled t(X) \%*\% y using the special structure of trend filtering}
\usage{
compute_tf_Xty(order, y)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{y}{an n vector}
}
\value{
an n vector
}
\description{
Compute unscaled t(X) \%*\% y using the special structure of trend filtering
}
\keyword{internal}
