% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfg_matrix.R
\name{compute_tfg_cm}
\alias{compute_tfg_cm}
\title{Compute column mean of the general trend filtering matrix X}
\usage{
compute_tfg_cm(X)
}
\arguments{
\item{X}{a general trend filtering matrix}
}
\value{
a p vector of column means
}
\description{
Compute column mean of the general trend filtering matrix X
}
\keyword{internal}
