% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfg_matrix.R
\name{compute_tfg_d}
\alias{compute_tfg_d}
\title{Compute d=diag(Xcent'Xcent) where Xcent is the centered version of the general trend filtering matrix X}
\usage{
compute_tfg_d(X)
}
\arguments{
\item{X}{a general trend filtering matrix}
}
\value{
a p vector d=diag(Xcent'Xcent)
}
\description{
Compute d=diag(Xcent'Xcent) where Xcent is the centered version of the general trend filtering matrix X
}
\keyword{internal}
