% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfg_matrix.R
\name{compute_tfg_Xb}
\alias{compute_tfg_Xb}
\title{Compute unscaled X \%*\% b using the special structure of trend filtering}
\usage{
compute_tfg_Xb(X, b)
}
\arguments{
\item{X}{a tfg_matrix created by make_tfg_matrix}

\item{b}{a p vector of the changes at each change point}
}
\value{
an n vector of the means at each data point
}
\description{
Compute unscaled X \%*\% b using the special structure of trend filtering
}
\keyword{internal}
