% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfg_matrix.R
\name{compute_tfg_Xty}
\alias{compute_tfg_Xty}
\title{Compute unscaled t(X) \%*\% y using the special structure of trend filtering}
\usage{
compute_tfg_Xty(X, y)
}
\arguments{
\item{X}{a tfg_matrix created by make_tfg_matrix}

\item{y}{an n vector of data}
}
\value{
a p vector
}
\description{
Compute unscaled t(X) \%*\% y using the special structure of trend filtering
}
\keyword{internal}
