% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_multiplication.R
\name{compute_Xb}
\alias{compute_Xb}
\title{Compute scaled.X \%*\% b using sparse multiplication}
\usage{
compute_Xb(X, b)
}
\arguments{
\item{X}{is a scaled dense matrix or an unscaled sparse matrix}

\item{b}{a p vector}
}
\value{
an n vector
}
\description{
Compute scaled.X \%*\% b using sparse multiplication
}
