% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_multiplication.R
\name{compute_Xty}
\alias{compute_Xty}
\title{Compute t(scaled.X)\%*\%y using sparse multiplication}
\usage{
compute_Xty(X, y)
}
\arguments{
\item{X}{is a scaled dense matrix or an unscaled sparse matrix}

\item{y}{an n vector}
}
\value{
a p vector
}
