% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{create_ash_grid}
\alias{create_ash_grid}
\title{Create variance grid for mr.ash}
\usage{
create_ash_grid(data, grid_length = 25, exponent = 2)
}
\description{
Create variance grid for mr.ash
}
\keyword{internal}
