% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_filtering_multiplication.R
\name{create_Dinv}
\alias{create_Dinv}
\title{Create D-inverse matrix as input X}
\usage{
create_Dinv(order, n)
}
\arguments{
\item{order}{is the order of trend filtering}

\item{n}{the length of y}
}
\value{
a D-inverse matrix
}
\description{
Create D-inverse matrix as input X
}
\keyword{internal}
