% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{data_small}
\alias{data_small}
\title{Simulated Small-sample eQTL Data.}
\format{
\code{data_small} is a list with the following elements:

\describe{

  \item{y}{Simulated gene expression response.}

  \item{X}{Genotype matrix.}}
}
\description{
A simulated eQTL data set with 47 individuals and 7,430
  variables. The response is a simulated gene expression phenotype and
  the variables are genotypes. This data set illustrates the small
  sample-size setting considered in Denault \emph{et al} (2025).
}
\examples{
data(data_small)
}
\references{
W. R. P. Denault \emph{et al} (2025). Accounting for uncertainty in
  residual variances improves fine-mapping in small sample studies.
  \emph{bioRxiv} \doi{10.1101/2025.05.16.654543}.
}
\seealso{
The \dQuote{Small data example} vignette.
}
\keyword{data}
