% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_reports.R
\name{diagnose_susie_ash_iter}
\alias{diagnose_susie_ash_iter}
\title{Diagnostic function for SuSiE-ash iteration}
\usage{
diagnose_susie_ash_iter(
  data,
  model,
  Xcorr,
  mrash_output,
  effect_purity,
  sentinels,
  current_collision,
  alpha_protected,
  pip_protected,
  neighborhood_pip,
  masked,
  sigma2_new,
  tau2_new,
  want_masked,
  ready_to_unmask,
  force_unmask,
  force_mask,
  purity_threshold,
  cs_formation_threshold,
  collision_ld_threshold,
  tight_ld_threshold,
  ld_threshold,
  diffuse_iter_count_param
)
}
\description{
Diagnostic function for SuSiE-ash iteration
}
\keyword{internal}
