% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_trendfilter.R
\name{estimate_mad_residual_variance}
\alias{estimate_mad_residual_variance}
\title{estimate residual variance using MAD estimator}
\usage{
estimate_mad_residual_variance(y)
}
\arguments{
\item{y}{an n vector}
}
\value{
a scalar of estimated residual variance
}
\description{
estimate residual variance using MAD estimator
}
\keyword{internal}
