% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residual_variance.R
\name{estimate_residual_variance}
\alias{estimate_residual_variance}
\title{Estimate residual variance}
\usage{
estimate_residual_variance(X, X.sparse, Y, cm, csd, s)
}
\arguments{
\item{X}{an n by p matrix of covariates, scaled}

\item{X.sparse}{an n by p matrix of covariates, unscaled}

\item{Y}{an n vector of regression outcome}

\item{cm}{a p vector of column means}

\item{csd}{a p vector of column standard deviations}

\item{s}{a susie fit}
}
\description{
Estimate residual variance
}
