% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residual_variance.R
\name{estimate_residual_variance_ss}
\alias{estimate_residual_variance_ss}
\title{Estimate residual variance for summary statistics}
\usage{
estimate_residual_variance_ss(XtX, Xty, s, var_y, n)
}
\arguments{
\item{XtX}{a p by p matrix}

\item{Xty}{a p vector}

\item{s}{a susie fit}

\item{var_y}{variance of y}

\item{n}{sample size}
}
\description{
Estimate residual variance for summary statistics
}
