% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residual_variance.R
\name{estimate_residual_variance_z}
\alias{estimate_residual_variance_z}
\title{Estimate residual variance for summary statistics}
\usage{
estimate_residual_variance_z(R, z, s)
}
\arguments{
\item{R}{a p by p}

\item{z}{a p vector}

\item{s}{a susie fit}
}
\description{
Estimate residual variance for summary statistics
}
