% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FinemappingConvergence}
\alias{FinemappingConvergence}
\title{Simulated Fine-mapping Data with Convergence Problem.}
\format{
\code{FinemappingConvergence} is a list with the following
elements:

\describe{

  \item{XtX}{XtX is computed using centered and scaled genotype
data.}

  \item{Xty}{XtX is computed using centered and scaled genotype
data and centered simulated response.}

  \item{yty}{yty is computed using centered simulated response.}

  \item{n}{Sample size.}

  \item{true_coef}{The simulated effect sizes.}

  \item{z}{Marginal z scores from simple linear regression.}}
}
\description{
The data is simulated using real genotypes
from 50000 individuals and 1001 SNPs. Two of the variables have
non-zero effects on the multivariate response. The response data
are generated under a linear regression model.
The genotypes and simulated response are column centered.
}
\examples{
data(FinemappingConvergence)
}
\keyword{data}
