% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.ash.R
\name{get.full.posterior}
\alias{get.full.posterior}
\title{Approximation Posterior Expectations from Mr.ASH Fit}
\usage{
get.full.posterior(fit)
}
\arguments{
\item{fit}{A Mr.ASH fit obtained, for example, by running
\code{mr.ash}.}
}
\value{
A list object with the following elements:

\item{phi}{An p x K matrix containing the posterior assignment
  probabilities, where p is the number of predictors, and K is the
  number of mixture components. (Each row of \code{phi} should sum to
  1.)}

\item{m}{An p x K matrix containing the posterior means conditional
  on assignment to each mixture component.}

\item{s2}{An p x K matrix containing the posterior variances
  conditional on assignment to each mixture component.}
}
\description{
Recover the parameters specifying the variational
  approximation to the posterior distribution of the regression
  coefficients. To streamline the model fitting implementation, and
  to reduce memory requirements, \code{\link{mr.ash}} does not store
  all the parameters needed to specify the approximate posterior.
}
\examples{
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

## fit mr.ash
fit.mr.ash  = mr.ash(X, y)

## recover full posterior
full.post   = get.full.posterior(fit.mr.ash)

}
