% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{get_alpha_l}
\alias{get_alpha_l}
\title{Get posterior inclusion probabilities for effect l}
\usage{
get_alpha_l(model, l)
}
\description{
Get posterior inclusion probabilities for effect l
}
\keyword{internal}
