% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{get_objective}
\alias{get_objective}
\title{Compute the SuSiE ELBO (evidence lower bound)}
\usage{
get_objective(data, params, model)
}
\arguments{
\item{data}{Data object.}

\item{params}{Params object.}

\item{model}{Model object.}
}
\value{
Scalar ELBO value.
}
\description{
Building-block function used by downstream packages implementing
custom IBSS loops.
}
\keyword{internal}
