% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_rss.R
\name{get_objective_rss}
\alias{get_objective_rss}
\title{Get objective function from data and susie fit object.}
\usage{
get_objective_rss(R, z, s)
}
\arguments{
\item{R}{p by p corelation matrix}

\item{z}{length p vector}

\item{s}{a susie fit object}
}
\description{
Get objective function from data and susie fit object.
}
\keyword{internal}
