% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_z.R
\name{get_objective_z}
\alias{get_objective_z}
\title{Get objective function from data and susie fit object.}
\usage{
get_objective_z(R, z, s)
}
\arguments{
\item{s}{a susie fit object}

\item{XtX}{a p by p matrix, X'X}

\item{Xty}{a p vector, X'Y,}
}
\description{
Get objective function from data and susie fit object.
}
\keyword{internal}
