% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{get_posterior_mean_l}
\alias{get_posterior_mean_l}
\title{Get PIP-weighted posterior mean for effect l (alpha * mu)}
\usage{
get_posterior_mean_l(model, l)
}
\description{
Get PIP-weighted posterior mean for effect l (alpha * mu)
}
\keyword{internal}
