% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{get_posterior_mean_sum}
\alias{get_posterior_mean_sum}
\title{Get sum of PIP-weighted posterior means across all effects}
\usage{
get_posterior_mean_sum(model)
}
\description{
Get sum of PIP-weighted posterior means across all effects
}
\keyword{internal}
