% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{get_posterior_moments_l}
\alias{get_posterior_moments_l}
\title{Get posterior moments for effect l (for EM prior variance update)}
\usage{
get_posterior_moments_l(model, l)
}
\description{
Get posterior moments for effect l (for EM prior variance update)
}
\keyword{internal}
