% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_finalize}
\alias{init_finalize}
\title{Update a susie fit object in order to initialize susie model.}
\usage{
init_finalize(s, X = NULL, Xr = NULL)
}
\description{
Update a susie fit object in order to initialize susie model.
}
