% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_setup}
\alias{init_setup}
\title{Set default susie initialization}
\usage{
init_setup(n, p, L, scaled_prior_variance, residual_variance,
  prior_weights, varY)
}
\description{
Set default susie initialization
}
