% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Check for symmetric matrix}
\usage{
is_symmetric_matrix(x)
}
\arguments{
\item{x}{A matrix to check}
}
\value{
Logical indicating if x is symmetric
}
\description{
Check for symmetric matrix
}
\keyword{internal}
