% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_effect_regression.R
\name{loglik}
\alias{loglik}
\title{The log likelihood function for SER model (based on summary data betahat, shat2), as a function of prior variance V}
\usage{
loglik(V, betahat, shat2, prior_weights)
}
\description{
The log likelihood function for SER model (based on summary data betahat, shat2), as a function of prior variance V
}
