% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N2finemapping}
\alias{N2finemapping}
\title{Add Title Here.}
\format{
\code{N2finemapping} is a list with the following elements:

\describe{
  \item{X}{Describe "X" here.}

  \item{chrom}{Describe "chrom" here.}

  \item{pos}{Describe "pos" here.}

  \item{true_coef}{Describe "true_coef" here.}

  \item{residual_variance}{Describe "residual_variance" here.}

  \item{Y}{Describe "Y" here.}

  \item{allele_freq}{Describe "allele_freq" here.}

  \item{V}{Describe "V" here.}
}
}
\description{
Briefly describe the data set here.
}
\keyword{data}
