% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N3finemapping}
\alias{N3finemapping}
\title{Add Title Here.}
\format{
\code{N3finemapping} is a list with the following elements:

\describe{
  \item{data$X}{Describe "X" here.}

  \item{data$chrom}{Describe "chrom" here.}

  \item{data$pos}{Describe "pos" here.}

  \item{data$true_coef}{Describe "true_coef" here.}

  \item{data$residual_variance}{Describe "residual_variance" here.}

  \item{Y}{Describe "Y" here.}

  \item{data$allele_freq}{Describe "allele_freq" here.}

  \item{data$V}{Describe "V" here.}

  \item{sumstats}{Describe "sumstats" here.}
}
}
\description{
Briefly describe the data set here.
}
\keyword{data}
