% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.ash.R
\name{predict.mr.ash}
\alias{predict.mr.ash}
\title{Predict Outcomes or Extract Coefficients from Mr.ASH Fit}
\usage{
\method{predict}{mr.ash}(object, newx = NULL, type = c("response", "coefficients"), ...)
}
\arguments{
\item{object}{A mr_ash fit, usually the result of calling
\code{mr.ash}.}

\item{newx}{The input matrix, of dimension (n,p); each column is a
single predictor; and each row is an observation vector. Here, n is
the number of samples and p is the number of predictors. When
\code{newx} is \code{NULL}, the fitted values for the training data
are provided.}

\item{type}{The type of output. For \code{type = "response"},
predicted or fitted outcomes are returned; for \code{type =
"coefficients"}, the estimated coefficients are returned.}

\item{...}{Additional arguments passed to the default S3 method.}
}
\value{
For \code{type = "response"}, predicted or fitted outcomes
are returned; for \code{type = "coefficients"}, the estimated
coefficients are returned.
}
\description{
This function predicts outcomes (y) given the observed
  variables (X) and a Mr.ASH model; alternatively, retrieve the
  estimates of the regression coefficients.
}
\examples{
## generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

## fit mr.ash
fit.mr.ash  = mr.ash(X, y)

## predict
Xnew        = matrix(rnorm(n*p),n,p)
ypred       = predict(fit.mr.ash, Xnew)

}
