% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.susie.R
\name{print.summary.susie}
\alias{print.summary.susie}
\title{Print summary(susie) result.}
\usage{
\method{print}{summary.susie}(object, ...)
}
\arguments{
\item{object}{A susie summary.}

\item{\dots}{Additional arguments passed to the generic \code{print}
method.}
}
\value{
standard output stream containing only variables in CS.
}
\description{
Print summary(susie) result.
}
