% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{safe_cor}
\alias{safe_cor}
\title{Compute a correlation matrix}
\usage{
safe_cor(X)
}
\arguments{
\item{X}{Data matrix (n x p)}
}
\value{
Correlation matrix (p x p)
}
\description{
Uses Rfast for nonconstant columns. Constant columns are handled by a
crossprod-based calculation and assigned zero correlation.
}
\keyword{internal}
