% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{SER_posterior_e_loglik}
\alias{SER_posterior_e_loglik}
\title{posterior expected loglikelihood for a single effect regression}
\usage{
SER_posterior_e_loglik(X, X.sparse, Y, cm, csd, s2, Eb, Eb2)
}
\arguments{
\item{X}{an n by p matrix of covariates, scaled}

\item{X.sparse}{an n by p matrix of covariates, unscaled}

\item{Y}{an n vector of regression outcome}

\item{cm}{a p vector of column means}

\item{csd}{a p vector of column standard deviations}

\item{s2}{the residual variance}

\item{Eb}{the posterior mean of b (p vector) (alpha * mu)}

\item{Eb2}{the posterior second moment of b (p vector) (alpha * mu2)}
}
\description{
posterior expected loglikelihood for a single effect regression
}
