% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_rss.R
\name{SER_posterior_e_loglik_rss}
\alias{SER_posterior_e_loglik_rss}
\title{posterior expected loglikelihood for a single effect regression}
\usage{
SER_posterior_e_loglik_rss(R, r, s2, Ez, Ez2)
}
\arguments{
\item{R}{a p by p symmetric and positive semidefinite correlation matrix.}

\item{r}{residuals}

\item{s2}{residual variance}

\item{Eb}{the posterior mean of b (p vector) (alpha * mu)}

\item{Eb2}{the posterior second moment of b (p vector) (alpha * mu2)}
}
\description{
posterior expected loglikelihood for a single effect regression
}
\keyword{internal}
