% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{set_prior_variance_l}
\alias{set_prior_variance_l}
\title{Set prior variance for effect l}
\usage{
set_prior_variance_l(model, l, V)
}
\description{
Set prior variance for effect l
}
\keyword{internal}
