% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_X_attributes.R
\name{set_X_attributes}
\alias{set_X_attributes}
\title{sets the attributes for the X matrix}
\usage{
set_X_attributes(X, center = TRUE, scale = TRUE)
}
\arguments{
\item{X}{an n by p data matrix that can be a dense, sparse, or trend filtering matrix}

\item{center}{boolean indicating mean centering or not}

\item{scale}{boolean indicating scaled by standard deviation or not}
}
\value{
X with three attributes e.g.
        attr(X, 'scaled:center') is a p vector of column means of X
        attr(X, 'scaled:scale') is a p vector of column standard deviations of X
        attr(X, 'd') is a p vector of column sums of X^2
if X is a trend filtering matrix, return the initial X with three attributes;
if X is a dense matrix, return the scaled X with three attributes;
if X is a sparse matrix, return the initial X with three attributes.
}
\description{
sets the attributes for the X matrix
}
