% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_effect_regression_ss.R
\name{single_effect_regression_ss}
\alias{single_effect_regression_ss}
\title{Bayesian single-effect linear regression of y on X using summary data}
\usage{
single_effect_regression_ss(
  Xty,
  dXtX,
  V = 1,
  residual_variance = 1,
  prior_weights = NULL,
  optimize_V = c("none", "optim", "uniroot", "EM", "simple"),
  check_null_threshold = 0
)
}
\arguments{
\item{Xty}{A p vector.}

\item{dXtX}{A p vector containing the diagonal elements of
\code{t(X) \%*\% X}.}

\item{V}{The prior variance.}

\item{residual_variance}{The residual variance.}

\item{prior_weights}{A p vector of prior weights.}

\item{optimize_V}{Specifies the method to optimize V (by maximum
likelihood).}

\item{check_null_threshold}{A sccalar threshold on the log-scale to
compare the likelihood between current estimate and zero (the
null).}
}
\value{
A list with the following elements:

\item{alpha}{Vector of posterior inclusion probabilities. i.e.,
  \code{alpha[i]} is posterior probability that that the ith
  coefficient is non-zero.}

\item{mu}{Vector of posterior means (conditional on inclusion).}

\item{mu2}{Vector of posterior second moments (conditional on
  inclusion).}

\item{lbf}{Vector of log-Bayes factors for each variable.}

\item{V}{The prior variance (after optimization if
  \code{optimize_V != "none"}).}

\item{lbf_model}{The log-likelihood for the total model minus the
  log-likelihood for the null model.}
}
\description{
Performs single-effect linear regression of y on X, in
  which only the summary statistcs \eqn{X^T y} and diagonal elements
  of \eqn{X^T X} are available. That is, this function fits the
  regression model eqn{y = Xb + e}, where elements of e are
  i.i.d. \eqn{N(0,s2)}, and the b is a p-vector of effects to be
  estimated. The assumption is that b has exactly one non-zero
  element, with all elements equally likely to be non-zero. The prior
  on the non-zero element is \eqn{N(0,V)}.
}
\keyword{internal}
