% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_vector.R
\name{split_vector}
\alias{split_vector}
\title{splits a vector into a list of vectors, each containing a specified number
of vectors of the orignal matrix}
\usage{
split_vector(b, nvar)
}
\arguments{
\item{b}{vector to be split}

\item{nvar}{vector of integers saying how many columns of b to include in each}
}
\description{
splits a vector into a list of vectors, each containing a specified number
of vectors of the orignal matrix
}
\examples{
b = c(rep(1,3),rep(2,4),rep(3,3))
split_vector(b,c(3,4,3))
}
