% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.susie.R
\name{summary.susie}
\alias{summary.susie}
\title{Summarize susie fit.}
\usage{
\method{summary}{susie}(object, ...)
}
\arguments{
\item{object}{A susie fit.}

\item{\dots}{Additional arguments passed to the generic \code{summary}
method.}
}
\value{
A list containing a data frame of variables and a data
  frame of credible sets.
}
\description{
Summarize susie fit.
}
