% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss.R
\name{susie_bhat}
\alias{susie_bhat}
\title{Summary statistics version of SuSiE on betahat, the corresponding standard error, and correlation (or covariance) matrix}
\usage{
susie_bhat(bhat, shat, R, n, var_y = 1, ...)
}
\arguments{
\item{bhat}{a p vector of estimated effects.}

\item{shat}{a p vector of corresponding standard errors.}

\item{R}{a p by p symmetric and positive semidefinite matrix. It can be X'X, covariance matrix (X'X/(n-1)) or correlation matrix.}

\item{n}{sample size.}

\item{var_y}{the (sample) variance of y, defined as y'y/(n-1) . If it is unknown, the coefficients (returned from `coef`) are on the standardized X, y scale.}

\item{...}{further arguments to be passed to \code{\link{susie_ss}}}
}
\value{
a susie fit
}
\description{
Summary statistics version of SuSiE on betahat, the corresponding standard error, and correlation (or covariance) matrix
}
