% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_CS}
\alias{susie_get_CS}
\title{Extract confidence sets from fitted SuSiE model}
\usage{
susie_get_CS(fitted, X = NULL, Xcorr = NULL, coverage = 0.95,
  min_abs_corr = 0.5)
}
\arguments{
\item{fitted}{a susie fit, the output of `susieR::susie()`, or simply the posterior
inclusion probability matrix alpha.}

\item{X}{N by P matrix of variables.
When provided, correlation between variables will be computed and used to remove
confidence sets whose minimum correlation between variables is smaller than `min_abs_corr` (see below).}

\item{Xcorr}{P by P matrix of correlations between variables.
when provided, it will be used to remove confidence sets
whose minimum correlation between variables is smaller than `min_abs_corr` (see below).}

\item{coverage}{coverage of confident sets.}

\item{min_abs_corr}{minimum of absolute value of correlation allowed in a confidence set.
Default set to 0.5 to correspond to squared correlation of 0.25,
a commonly used threshold for genotype data in genetics studies.}
}
\value{
a list of `cs` and `purity`
}
\description{
Extract confidence sets from fitted SuSiE model
}
\details{
It reports indices of variables in each confidence set identified,
as well as summaries of correlation between variables within each set.
}
