% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_lfsr}
\alias{susie_get_lfsr}
\title{Local false sign rate (lfsr) for susie credible sets.}
\usage{
susie_get_lfsr(res)
}
\arguments{
\item{res}{A susie fit, typically an output from \code{\link{susie}}.}
}
\value{
An l-vector of lfsr for credible sets.
}
\description{
Computes the average lfsr across SNPs for each single l,
  weighted by the posterior inclusion probability, alpha.
}
