% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_niter}
\alias{susie_get_niter}
\title{Get number of iterations from fitted susie model}
\usage{
susie_get_niter(res)
}
\arguments{
\item{res}{a susie fit}
}
\description{
Get number of iterations from fitted susie model
}
