% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{susie_get_objective}
\alias{susie_get_objective}
\title{Get objective function from data and susie fit object.}
\usage{
susie_get_objective(X, Y, s)
}
\arguments{
\item{data}{A flash data object.}

\item{f}{A flash fit object.}
}
