% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_ss.R
\name{susie_get_objective_ss}
\alias{susie_get_objective_ss}
\title{Get objective function from data and susie fit object.}
\usage{
susie_get_objective_ss(XtX, Xty, s, var_y, n)
}
\arguments{
\item{XtX}{a p by p matrix, X'X}

\item{Xty}{a p vector, X'Y,}

\item{s}{a susie fit object}

\item{var_y}{the (sample) variance of the vector Y}

\item{n}{sample size}
}
