% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_posterior_mean}
\alias{susie_get_posterior_mean}
\title{Get posterior mean for coefficients from fitted SuSiE model}
\usage{
susie_get_posterior_mean(res)
}
\arguments{
\item{res}{a susie fit}
}
\description{
Get posterior mean for coefficients from fitted SuSiE model
}
