% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_posterior_sd}
\alias{susie_get_posterior_sd}
\title{Get posterior standard deviation for coefficients from fitted SuSiE model}
\usage{
susie_get_posterior_sd(res)
}
\arguments{
\item{res}{a susie fit}
}
\description{
Get posterior standard deviation for coefficients from fitted SuSiE model
}
