% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_prior_variance}
\alias{susie_get_prior_variance}
\title{Get updated prior variance}
\usage{
susie_get_prior_variance(res)
}
\arguments{
\item{res}{a susie fit, the output of `susieR::susie()`.}
}
