% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_in_CS}
\alias{susie_in_CS}
\title{Variables in susie confidence sets}
\usage{
susie_in_CS(res)
}
\arguments{
\item{fitted}{a susie fit, the output of `susieR::susie()`, or simply the posterior
inclusion probability matrix alpha}
}
\value{
a l by p binary matrix
}
\details{
It returns a binary matrix indicating which variables are in CS
of each effect
}
