% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_logistic.R
\name{susie_logistic}
\alias{susie_logistic}
\title{SUm of Single Effect (SuSiE) Logistic Regression of Y on X}
\usage{
susie_logistic(X, Y, L = min(10, ncol(X)), prior_variance = 1,
  prior_weights = NULL, intercept = TRUE, Z = NULL,
  estimate_prior_variance = FALSE, s_init = NULL, coverage = 0.95,
  min_abs_corr = 0.5, max_iter = 100, tol = 0.001, verbose = FALSE,
  track_fit = FALSE)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{Y}{an n vector (binary)}

\item{L}{maximum number of non-zero effects}

\item{prior_variance}{the prior variance (vector of length L, or scalar. In latter case gets repeated L times). The prior variance on each non-zero element of b is set to be prior_variance.}

\item{prior_weights}{a p vector of prior probability that each element is non-zero}

\item{intercept}{Should intercept be fitted (default=TRUE) or set to zero (FALSE). The latter is generally not recommended.}

\item{Z}{an n by q vector of covariates to include in the model (e.g. top 10 PCs, gender, etc). The estimated effects
for these covarites are not penalized, and there is no prior on the effects, the algorithm simply maximizes the ELBO w.r.t. the effects.
Do not include a constant column for the intercept, this is handled separately.}

\item{estimate_prior_variance}{indicates whether to estimate prior (initial estimate given by prior_variance. If prior_variance is
a vector, then a different prior variance is estimated for each of the L effects)}

\item{s_init}{a previous susie fit with which to initialize (NOT CURRENTLY IMPLEMENTD IN LOGISTIC VERSION)}

\item{coverage}{coverage of confident sets. Default to 0.95 for 95\% credible interval.}

\item{min_abs_corr}{minimum of absolute value of correlation allowed in a credible set.
Default set to 0.5 to correspond to squared correlation of 0.25,
a commonly used threshold for genotype data in genetics studies.}

\item{max_iter}{maximum number of iterations to perform}

\item{tol}{convergence tolerance}

\item{verbose}{if true outputs some progress messages}

\item{track_fit}{add an attribute \code{trace} to output that saves current values of all iterations (NOT CURRENTLY IMPLEMENTD IN LOGISTIC VERSION)}
}
\value{
a susie fit, which is a list with some or all of the following elements\cr
\item{alpha}{an L by p matrix of posterior inclusion probabilites}
\item{mu}{an L by p matrix of posterior means (conditional on inclusion)}
\item{mu2}{an L by p matrix of posterior second moments (conditional on inclusion)}
\item{Xr}{an n vector of fitted values, equal to X times colSums(alpha*mu))}
\item{sigma2}{residual variance}
\item{V}{prior variance}
\item{elbo}{a vector of values of elbo achieved (objective function)}
\item{sets}{a list of `cs`, `purity` and selected `cs_index`}
\item{pip}{a vector of posterior inclusion probability}
\item{z}{a vector of univariate z-scores}
}
\description{
SUm of Single Effect (SuSiE) Logistic Regression of Y on X
}
\details{
Performs Bayesian logistic regression of Y on X.
That is, this function
fits the regression model Y ~ Bern(sum_l Xb_l), where
sum_l b_l is a p vector of effects to be estimated.
The assumption is that each b_l has exactly one non-zero element, with all elements
equally likely to be non-zero. The prior on the non-zero element is N(0, var = prior_variance).
}
\examples{
set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1:4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = rbinom(1000, 1, 1 / (1 + exp(-X \%*\% beta)))
res = susie_logistic(X,y,L=10)
coef(res)
plot(y,predict(res))

}
