% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_plot_iteration}
\alias{susie_plot_iteration}
\title{Diagnostic plot for SuSiE iterations}
\usage{
susie_plot_iteration(model, L, file_prefix, pos = NULL)
}
\arguments{
\item{model}{a susie fit, the output of `susieR::susie()`.
Multiple plots will be made for all iterations if `track_fit` was set to `TRUE` when running SuSiE.}

\item{L}{an integer, number of CS to plot}

\item{file_prefix}{prefix to path of output plot file}

\item{pos}{index of variables to plot, default to all variables}
}
\description{
Diagnostic plot for SuSiE iterations
}
