% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie_set_init}
\alias{susie_set_init}
\title{initialize a susie object using given nonzero effect indices and beta values}
\usage{
susie_set_init(coef_index, coef_value, num_variables, prior_variance,
  residual_variance = NULL)
}
\arguments{
\item{coef_index}{a L-vector for indices of nonzero effects}

\item{coef_value}{a L-vector for initial estimated beta values}

\item{prior_variance}{a scalar or an L-vector}

\item{residual_variance}{a scalar}

\item{num_varialbes}{a scalar the number of variables in the data}
}
\value{
a list of initialized alpha, mu, mu2, sa2 and optionally sigma2
}
\description{
initialize a susie object using given nonzero effect indices and beta values
}
