% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_slim}
\alias{susie_slim}
\title{Slim the result of fitted SuSiE model}
\usage{
susie_slim(res)
}
\arguments{
\item{res}{a susie fit, the output of `susieR::susie()`}
}
\description{
Slim the result of fitted SuSiE model
}
