% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss_set_data.R
\name{susie_ss_set_data}
\alias{susie_ss_set_data}
\title{Create data for susie_ss}
\usage{
susie_ss_set_data(bhat = NULL, shat = NULL, zhat = NULL, XtX = NULL,
  R = NULL, var_y = 1, n, standardize = TRUE, bhat.standardized = TRUE)
}
\arguments{
\item{bhat}{a p vector, the estimated effects from p simple linear regression: y = Xb + e}

\item{shat}{a p vector of corresponding standard errors}

\item{zhat}{a p vector of z scores. bhat/shat}

\item{XtX}{a p by p matrix, X'X, where columns of X are centered to have mean 0}

\item{R}{a p by p correlation matrix of X}

\item{var_y}{the (sample) variance of y}

\item{n}{sample size}

\item{standardize}{logical flag (default=TRUE) for whether to fit susie_ss with standardize columns of X}

\item{bhat.standardized}{whether the bhat comes from the standardized X}
}
