% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_stumps.R
\name{susie_stumps}
\alias{susie_stumps}
\title{Applies susie to perform non-linear regression using "stumps" (trees
with  a  single split)}
\usage{
susie_stumps(X, y, ...)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{y}{an n vector of observations}

\item{...}{other parameters to pass to susie}
}
\value{
a "susie" object; see ?susie for details
}
\description{
Applies susie to perform non-linear regression using "stumps" (trees
with  a  single split)
}
\details{
TBD
}
\examples{
TBD
}
