% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_stumps.R
\name{susie_stumps_old}
\alias{susie_stumps_old}
\title{Applies susie to perform non-linear regression using "stumps" (trees
with  a  single split). Old version (original beta implementation)}
\usage{
susie_stumps_old(X, Y, ...)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{Y}{an n vector of observations}

\item{...}{other parameters to pass to susie}
}
\value{
a "susie" object; see ?susie for details
}
\description{
Applies susie to perform non-linear regression using "stumps" (trees
with  a  single split). Old version (original beta implementation)
}
\details{
TBD
}
\examples{
TBD
}
