% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_rss.R
\name{susie_z}
\alias{susie_z}
\title{SuSiE on z scores and correlation matrix.}
\usage{
susie_z(z, R, ...)
}
\arguments{
\item{z}{a p vector of z scores.}

\item{R}{a p by p symmetric and positive semidefinite correlation matrix.}

\item{...}{further arguments to be passed to \code{\link{susie_rss}}}
}
\value{
a susie fit
}
\description{
SuSiE on z scores and correlation matrix.
}
