% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_regression.R
\name{univar.order}
\alias{univar.order}
\title{Ordering of Predictors from Univariate Regression}
\usage{
univar.order(X, y)
}
\arguments{
\item{X}{An input design matrix. This may be centered and/or
standardized prior to calling function.}

\item{y}{A vector of response variables.}
}
\value{
An ordering of the predictors.
}
\description{
This function extracts the ordering of the predictors
  according to the coefficients estimated in a basic univariate
  regression; in particular, the predictors are ordered in decreasing
  order by magnitude of the univariate regression coefficient
  estimate.
}
\examples{
### generate synthetic data
set.seed(1)
n           = 200
p           = 300
X           = matrix(rnorm(n*p),n,p)
beta        = double(p)
beta[1:10]  = 1:10
y           = X \%*\% beta + rnorm(n)

univ.order = univar.order(X,y)

}
