% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{unmappable_data}
\alias{unmappable_data}
\title{Simulated Fine-mapping Data with Sparse, Oligogenic and Polygenic Effects.}
\format{
\code{unmappable_data} is a list with the following elements:

\describe{

  \item{X}{Centered and scaled genotype matrix.}

  \item{y}{Simulated response.}

  \item{beta}{Simulated effect sizes.}

  \item{h2g}{Total proportion of variance in the response explained
    by the simulated effects.}}
}
\description{
A simulated data set with 1,000 individuals and 5,000
  variants, combining 3 sparse, 5 oligogenic and 15 polygenic
  non-zero effects. The response is generated under a linear
  regression model. This data set illustrates fine-mapping with
  SuSiE-ash and SuSiE-inf.
}
\examples{
data(unmappable_data)
}
\seealso{
The \dQuote{Fine-mapping with SuSiE-ash and SuSiE-inf}
  vignette.
}
\keyword{data}
