% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_vbupdate.R
\name{update_each_effect}
\alias{update_each_effect}
\title{update each effect once}
\usage{
update_each_effect(X, Y, s_init)
}
\arguments{
\item{X}{an n by p matrix of covariantes}

\item{Y}{an n vector of data}

\item{s_init}{a list with elements sigma2, sa2, alpha, mu, Xr}
}
\description{
update each effect once
}
