% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_each_effect_rss.R
\name{update_each_effect_rss}
\alias{update_each_effect_rss}
\title{update each effect once}
\usage{
update_each_effect_rss(
  R,
  z,
  s,
  estimate_prior_variance = FALSE,
  estimate_prior_method = "optim"
)
}
\arguments{
\item{R}{a p by p symmetric and positive semidefinite correlation matrix.}

\item{z}{a p vector}

\item{s}{a list with elements sigma2, V, alpha, mu, Xr}

\item{estimate_prior_variance}{boolean indicating whether to estimate prior variance}
}
\description{
update each effect once
}
\keyword{internal}
