% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.susie.R
\name{coef.susie}
\alias{coef.susie}
\title{extract regression coefficients from susie fit}
\usage{
\method{coef}{susie}(object, ...)
}
\arguments{
\item{object}{a susie fit}
}
\value{
a p+1 vector, the first element being an intercept, and the remaining p elements being estimated regression coefficients
}
\description{
extract regression coefficients from susie fit
}
