% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_multiplication.R
\name{compute_MXt}
\alias{compute_MXt}
\title{Computes M\%*\%t(standardized.X) using sparse multiplication trick}
\usage{
compute_MXt(M, X)
}
\arguments{
\item{M}{a L by p matrix}

\item{X}{an n by p unstandardized matrix with three attributes: attr(X, 'scaled:center'), attr(X, 'scaled:scale'), and attr(X, 'd')}
}
\value{
a L by n matrix
}
\description{
Computes M\%*\%t(standardized.X) using sparse multiplication trick
}
