% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ss.R
\name{compute_ss}
\alias{compute_ss}
\title{Compute the needed summary statistics for `susie_ss` from the 'raw' data}
\usage{
compute_ss(X, y, standardize = TRUE)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{y}{an n vector}

\item{standardize}{logical flag (default=TRUE) for whether to standardize columns of X to unit variance prior to fitting.}
}
\value{
a list with the input for `susie_ss`
}
\description{
Compute the needed summary statistics for `susie_ss` from the 'raw' data
}
\keyword{internal}
