% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ss.R
\name{compute_ss}
\alias{compute_ss}
\title{Compute sufficient statistics for susie.}
\usage{
compute_ss(X, y, standardize = TRUE)
}
\arguments{
\item{X}{An n by p matrix of covariates.}

\item{y}{An n vector.}

\item{standardize}{Logical flag indicating whether to standardize
columns of X to unit variance prior to fitting.}
}
\value{
A list of sufficient statistics.
}
\description{
Compute sufficient statistics for susie.
}
