% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_multiplication.R
\name{compute_Xb}
\alias{compute_Xb}
\title{Computes standardized.X \%*\% b using sparse multiplication trick}
\usage{
compute_Xb(X, b)
}
\arguments{
\item{X}{an n by p unstandardized matrix with three attributes: attr(X, 'scaled:center'), attr(X, 'scaled:scale'), and attr(X, 'd')}

\item{b}{a p vector}
}
\value{
an n vector
}
\description{
Computes standardized.X \%*\% b using sparse multiplication trick
}
\keyword{internal}
