% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_multiplication.R
\name{compute_Xty}
\alias{compute_Xty}
\title{Computes t(standardized.X)\%*\%y using sparse multiplication trick}
\usage{
compute_Xty(X, y)
}
\arguments{
\item{X}{an n by p unstandardized matrix with three attributes: attr(X, 'scaled:center'), attr(X, 'scaled:scale'), and attr(X, 'd')}

\item{y}{an n vector}
}
\value{
a p vector
}
\description{
Computes t(standardized.X)\%*\%y using sparse multiplication trick
}
