% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residual_variance.R
\name{estimate_residual_variance}
\alias{estimate_residual_variance}
\title{Estimate residual variance}
\usage{
estimate_residual_variance(X, Y, s)
}
\arguments{
\item{X}{an n by p matrix of covariantes}

\item{Y}{an n vector of data}

\item{s}{a susie fit}
}
\description{
Estimate residual variance
}
