% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FinemappingConvergence}
\alias{FinemappingConvergence}
\title{Add Title Here}
\format{
\code{FinemappingConvergence} is a list with the following
elements:

\describe{

  \item{XtX}{Describe XtX here.}

  \item{Xty}{Describe Xty here.}

  \item{yty}{Describe yty here.}

  \item{n}{Describe n here.}

  \item{true_coef}{Describe true_coef here.}

  \item{z}{Describe z here.}}
}
\description{
Add description here.
}
\examples{
data(FinemappingConvergence)
}
\keyword{data}
