% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FinemappingConvergence}
\alias{FinemappingConvergence}
\title{Simulated Fine-mapping Data with Convergence Problem.}
\format{
\code{FinemappingConvergence} is a list with the following
elements:

\describe{

  \item{XtX}{Summary statistics computed using the centered and
    scaled genotype matrix.}

  \item{Xty}{Summary statistics computed using the centered and
    scaled genotype data, and the centered simulated response.}

  \item{yty}{yty is computed using the centered simulated response.}

  \item{n}{The sample size (50,000).}

  \item{true_coef}{The coefficients used to simulate the responses.}

  \item{z}{z-scores from a simple (single-SNP) linear regression.}}
}
\description{
Data simulated using real genotypes from 50,000
  individuals and 200 SNPs. Two of the SNPs have non-zero effects
  on the multivariate response. The response data are generated under
  a linear regression model. The simulated response and the columns
  of the genotype matrix are centered.
}
\examples{
data(FinemappingConvergence)
}
\seealso{
A similar data set with more SNPs is used in the
  \dQuote{Refine SuSiE model} vignette.
}
\keyword{data}
