% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{get_objective}
\alias{get_objective}
\title{Get objective function from data and susie fit object.}
\usage{
get_objective(X, Y, s)
}
\arguments{
\item{data}{A flash data object.}

\item{f}{A flash fit object.}
}
\description{
Get objective function from data and susie fit object.
}
\keyword{internal}
