% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_ss.R
\name{get_objective_ss}
\alias{get_objective_ss}
\title{Get objective function from data and susie fit object.}
\usage{
get_objective_ss(XtX, Xty, s, yty, n)
}
\arguments{
\item{XtX}{a p by p matrix, X'X}

\item{Xty}{a p vector, X'y,}

\item{s}{a susie fit object}

\item{yty}{a scaler, y'y, where y is centered to have mean 0}

\item{n}{sample size}
}
\description{
Get objective function from data and susie fit object.
}
\keyword{internal}
