% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfg_matrix.R
\name{make_tfg_matrix}
\alias{make_tfg_matrix}
\title{set up a general trend filtering matrix}
\usage{
make_tfg_matrix(t, br = t, order = 0)
}
\arguments{
\item{t}{vector of length n specifying locations of data points on x axis}

\item{br}{vector of length (p-1) specifying break points on x axis (ie where changepoints can occur)
By default br=t which allows breaks to occur between every data point. Note that internally duplicate elements of br are removed.}

\item{order}{non-negative integer indicating order of trend filtering basis (0 is changepoint basis and is the only case we test and use)}
}
\description{
set up a general trend filtering matrix
}
\keyword{internal}
