% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.susie.R
\name{predict.susie}
\alias{predict.susie}
\title{Predict outcomes or extract coefficients from susie fit.}
\usage{
\method{predict}{susie}(object, newx = NULL, type = c("response", "coefficients"), ...)
}
\arguments{
\item{object}{A susie fit.}

\item{newx}{A new value for X at which to do predictions.}

\item{type}{The type of output. For \code{type = "response"},
predicted or fitted outcomes are returned; for \code{type =
"coefficients"}, the estimated coefficients are returned.}

\item{\dots}{Other arguments used by generic predict function. These
extra arguments are not used here.}
}
\value{
For \code{type = "response"}, predicted or fitted outcomes
  are returned; for \code{type = "coefficients"}, the estimated
  coefficients are returned. The prediction ignores the intercept (intercept = 0)
  if there is no intercept in susie fit.
}
\description{
Predict outcomes or extract coefficients from susie fit.
}
