% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.susie.R
\name{predict.susie}
\alias{predict.susie}
\title{predict future observations or extract coefficients from susie fit}
\usage{
\method{predict}{susie}(object, newx = NULL, type = c("response",
  "coefficients"), ...)
}
\arguments{
\item{object}{a susie fit}

\item{newx}{a new value for X at which to do predictions}

\item{type}{if this is coefficients, then calls coef.susie}
}
\description{
predict future observations or extract coefficients from susie fit
}
