% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_R_attributes.R
\name{set_R_attributes}
\alias{set_R_attributes}
\title{sets the attributes for the R matrix}
\usage{
set_R_attributes(R, r_tol)
}
\arguments{
\item{R}{a p by p LD matrix}

\item{r_tol}{tolerance level for eigen value check of positive semidefinite matrix of R}
}
\value{
R with attribute e.g.
        attr(R, 'eigen') is the eigen decomposition of R.
}
\description{
sets the attributes for the R matrix
}
