% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_R_attributes.R
\name{set_R_attributes}
\alias{set_R_attributes}
\title{sets the attributes for the R matrix}
\usage{
set_R_attributes(R, expected_dim)
}
\arguments{
\item{R}{a p by p LD matrix}

\item{z}{a p vector of z scores}
}
\value{
R with attribute e.g.
        attr(R, 'eigenR') is the eigen decomposition of R.
}
\description{
sets the attributes for the R matrix
}
