% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_effect_regression.R
\name{single_effect_regression}
\alias{single_effect_regression}
\title{Bayesian single-effect linear regression of y on X}
\usage{
single_effect_regression(
  Y,
  X,
  V,
  residual_variance = 1,
  prior_weights = NULL,
  optimize_V = c("none", "optim", "uniroot", "EM", "simple"),
  check_null_threshold = 0
)
}
\arguments{
\item{Y}{An n vector.}

\item{X}{An n by p matrix of covariates.}

\item{V}{The prior variance.}

\item{residual_variance}{The residual variance.}

\item{prior_weights}{A p vector of prior weights.}

\item{optimize_V}{Boolean indicating whether to optimize V (by
maximum likelihood).}

\item{check_null_threshold}{Scalar specifying threshold on the
log-scale to compare likelihood between current estimate and zero
the null.}
}
\value{
A list with the following elements:

\item{alpha}{Vector of posterior inclusion probabilities;
  \code{alpha[i]} is posterior probability that that the ith
  coefficient is non-zero}

\item{mu}{Vector of posterior means (conditional on inclusion).}

\item{mu2}{Vector of posterior second moments (conditional on
  inclusion).}

\item{lbf}{Vector of log-Bayes factors for each variable.}

\item{lbf_model}{Log-Bayes factor for the single effect regression.}

\item{V}{Prior variance (after optimization if \code{optimize_V = TRUE}).}

\item{loglik}{The logarithm of the likelihood, \eqn{p(y | X, V)}.}
}
\description{
Performs single-effect linear regression of y on
  X. That is, this function fits the regression model \eqn{y = Xb +
  e}, where elements of e are i.i.d. \eqn{N(0,s2)}, and b is a
  p-vector of effects to be estimated. The assumption is that b has
  exactly one non-zero element, with all elements equally likely to
  be non-zero. The prior on the coefficient of the non-zero element
  is \eqn{N(0,V)}.
}
\keyword{internal}
