% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_effect_regression_ss.R
\name{single_effect_regression_ss}
\alias{single_effect_regression_ss}
\title{Bayesian single-effect linear regression of Y on X}
\usage{
single_effect_regression_ss(Xty, dXtX, V = 1, residual_variance = 1,
  optimize_V = FALSE)
}
\arguments{
\item{Xty}{an p vector}

\item{dXtX}{an p vector, diagonal elements of t(X)X}

\item{V}{the prior variance}

\item{residual_variance}{the residual variance}

\item{optimize_V}{boolean indicating whether to optimize V (by maximum likelihood)}
}
\value{
a list with elements: \cr
\item{alpha}{vector of posterior inclusion probabilities. ie alpha[i] is posterior probability that
 that b[i] is non-zero}
\item{mu}{vector of posterior means (conditional on inclusion)}
\item{mu2}{vector of posterior second moments (conditional on inclusion)}
\item{lbf}{vector of log Bayes factors for each variable}
\item{V}{the prior variance (after optimization, if optimize_V is TRUE)}
\item{logBF}{(scalar) the loglikelihood for the total model minus the log-likelihood for the null model}
}
\description{
Bayesian single-effect linear regression of Y on X
}
\details{
Performs single-effect linear regression of Y on X. That is, this function
fits the regression model Y= Xb + e, where elements of e are iid N(0,residual_variance) and the
b is a p vector of effects to be estimated.
The assumption is that b has exactly one non-zero element, with all elements
equally likely to be non-zero. The prior on the non-zero element is N(0,var=V).
Only the summary statistcs t(X)Y and diagonal elements of t(X)X are avialable.
}
