% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.susie.R
\name{summary.susie}
\alias{summary.susie}
\title{summarize susie fit}
\usage{
\method{summary}{susie}(object, ...)
}
\arguments{
\item{object}{a susie fit}
}
\value{
a data.frame of variables and a data.frame of credible sets
}
\description{
summarize susie fit
}
