% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SummaryConsistency}
\alias{SummaryConsistency}
\title{Simulated Fine-mapping Data with LD matrix From Reference Panel.}
\format{
\code{SummaryConsistency} is a list with the following
elements:

\describe{

  \item{z}{z-scores computed by fitting univariate simple regression
    variable-by-variable.}

  \item{ldref}{LD matrix estimated from the reference panel.}

  \item{flip_id}{The index of the SNP with the flipped allele.}

  \item{signal_id}{The index of the SNP with the non-zero effect.}}
}
\description{
Data simulated using real genotypes from 10,000
  individuals and 200 SNPs. One SNP have non-zero effect
  on the multivariate response. The response data are generated under
  a linear regression model. There is also one SNP with flipped allele
  between summary statistics and the reference panel.
}
\examples{
data(SummaryConsistency)
}
\seealso{
A similar data set with more samples is used in the
  \dQuote{Diagnostic for fine-mapping with summary statistics}
  vignette.
}
\keyword{data}
