% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie}
\alias{susie}
\title{Bayesian sum of single-effect (susie) linear regression of Y on X}
\usage{
susie(X, Y, sa2 = 1, sigma2 = NULL, niter = 100, L = 5,
  calc_elbo = FALSE)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{Y}{an n vector}

\item{sa2}{the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times )}

\item{sigma2}{the residual variance (defaults to variance of Y)}

\item{niter}{number of iterations of vb method}

\item{L}{number of single effects}

\item{calc_elbo}{indicates whether to compute the evidence lower bound (could slow things down; useful for testing)}
}
\value{
a susie fit, which is a list with some or all of the following elements\cr
\item{alpha}{an L by p matrix of posterior inclusion probabilites}
\item{mu}{an L by p matrix of posterior means}
\item{postvar}{an L by p matrix of posterior variances}
\item{Xr}{an n vector of fitted values, equal to X times colSums(alpha*mu))}
\item{sigma2}{residual variance}
\item{sa2}{scaled prior variance}
\item{elbo}{vector of values of elbo achieved (if calc_elbo is TRUE)}
}
\description{
Bayesian sum of single-effect (susie) linear regression of Y on X
}
\details{
Performs sum of single-effect (susie) linear regression of Y on X.
That is, this function
fits the regression model Y= sum_l Xb_l + e, where elements of e are iid N(0,s2) and the
sum_l b_l is a p vector of effects to be estimated.
The assumption is that each b_l has exactly one non-zero element, with all elements
equally likely to be non-zero. The prior on the non-zero element is N(0,var=sa2*s2).
}
\examples{
set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1] = 1
beta[2] = 1
beta[3] = 1
beta[4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = X \%*\% beta + rnorm(n)
res =susie(X,y,niter=10,L=5,calc_elbo = TRUE)
coef(res)
plot(y,predict(res))
}
