% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie}
\alias{susie}
\title{Bayesian sum of single-effect (susie) linear regression of Y on X}
\usage{
susie(X, Y, L = 10, prior_variance = 1, residual_variance = NULL,
  max_iter = 100, tol = 0.01, estimate_residual_variance = TRUE,
  estimate_prior_variance = FALSE, s_init = NULL, verbose = FALSE)
}
\arguments{
\item{X}{an n by p matrix of covariates}

\item{Y}{an n vector}

\item{L}{maximum number of non-zero effects}

\item{prior_variance}{the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times )}

\item{residual_variance}{the residual variance (defaults to variance of Y)}

\item{max_iter}{maximum number of iterations to perform}

\item{tol}{convergence tolerance}

\item{estimate_residual_variance}{indicates whether to estimate residual variance}

\item{estimate_prior_variance}{indicates whether to estimate prior (currently not recommended as not working as well)}

\item{s_init}{a previous susie fit with which to initialize}

\item{verbose}{if true outputs some progress messages}
}
\value{
a susie fit, which is a list with some or all of the following elements\cr
\item{alpha}{an L by p matrix of posterior inclusion probabilites}
\item{mu}{an L by p matrix of posterior means (conditional on inclusion)}
\item{mu2}{an L by p matrix of posterior second moments (conditional on inclusion)}
\item{Xr}{an n vector of fitted values, equal to X times colSums(alpha*mu))}
\item{sigma2}{residual variance}
\item{sa2}{scaled prior variance; ie prior variance is sigma2*sa2}
\item{elbo}{vector of values of elbo achieved (objective function)}
}
\description{
Bayesian sum of single-effect (susie) linear regression of Y on X
}
\details{
Performs sum of single-effect (susie) linear regression of Y on X.
That is, this function
fits the regression model Y= sum_l Xb_l + e, where elements of e are iid N(0,s2) and the
sum_l b_l is a p vector of effects to be estimated.
The assumption is that each b_l has exactly one non-zero element, with all elements
equally likely to be non-zero. The prior on the non-zero element is N(0,var=sa2*s2).
}
\examples{
set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1] = 1
beta[2] = 1
beta[3] = 1
beta[4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = X \%*\% beta + rnorm(n)
res =susie(X,y,L=10)
coef(res)
plot(y,predict(res))
}
