% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_auto.R
\name{susie_auto}
\alias{susie_auto}
\title{An attempt to automate reliable running of susie even on
  hard problems}
\usage{
susie_auto(
  X,
  Y,
  L_init = 1,
  L_max = 512,
  verbose = FALSE,
  init_tol = 1,
  standardize = TRUE,
  intercept = TRUE,
  max_iter = 100,
  tol = 0.01,
  ...
)
}
\arguments{
\item{X}{An n by p matrix of covariates.}

\item{Y}{An n vector.}

\item{L_init}{The initial value of L.}

\item{L_max}{The maximum value of L to consider.}

\item{verbose}{If \code{TRUE}, some progress messages are outputted.}

\item{init_tol}{The tolerance to passed to susie during early runs
(set large to shorten the initial runs).}

\item{standardize}{Logical; whether or not to standardize columns
of X to unit variance prior to fitting. See \code{\link{susie}} for
details.}

\item{intercept}{Logical; should intercept be fitted (\code{TRUE})
or fixed to zero (\code{FALSE}).}

\item{max_iter}{Maximum number of iterations to perform.}

\item{tol}{Convergence tolerance.}

\item{\dots}{Additional arguments passed to \code{\link{susie}}.}
}
\description{
Implements a three-stage strategy for each L: first
  fit susie with very small residual error; next, estimate residual
  error; finally, estimate the prior variance. If the last step
  estimates some prior variances to be zero, stop. Otherwise, double
  L, and repeat.  Initial runs are performed with relaxed tolerance;
  the final run is performed using the default susie tolerance.
}
\seealso{
\code{\link{susie}}
}
