% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_lfsr}
\alias{susie_get_lfsr}
\title{Local false sign rate (lfsr) for susie confidence sets}
\usage{
susie_get_lfsr(res)
}
\arguments{
\item{fitted}{a susie fit, the output of `susieR::susie()`}
}
\value{
an l vector of lfsr for confidence sets
}
\description{
Local false sign rate (lfsr) for susie confidence sets
}
\details{
This computes the average lfsr across SNPs for each l, weighted by the
posterior inclusion probability alpha
}
