% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_objective}
\alias{susie_get_objective}
\alias{susie_get_posterior_mean}
\alias{susie_get_posterior_sd}
\alias{susie_get_niter}
\alias{susie_get_prior_variance}
\alias{susie_get_residual_variance}
\alias{susie_get_lfsr}
\alias{susie_get_cs}
\alias{susie_get_pip}
\title{Add Title Here.}
\usage{
susie_get_objective(res, last_only = TRUE, warning_tol = 1e-06)

susie_get_posterior_mean(res)

susie_get_posterior_sd(res)

susie_get_niter(res)

susie_get_prior_variance(res)

susie_get_residual_variance(res)

susie_get_lfsr(res)

susie_get_cs(
  res,
  X = NULL,
  Xcorr = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  dedup = TRUE,
  squared = FALSE
)

susie_get_pip(res, prune_by_cs = FALSE, prior_tol = 1e-09)
}
\arguments{
\item{res}{a susie fit, the output of `susieR::susie()`, or simply
the posterior inclusion probability matrix alpha.}

\item{last_only}{Whether or not to get ELBO from all iterations.}

\item{warning_tol}{Warn if ELBO is non-decreasing by this tolerance
level.}

\item{X}{n by p matrix of values of the P variables (covariates) in
n samples.  When provided, correlation between variables will be
computed and used to remove CSs whose minimum correlation among
variables is smaller than `min_abs_corr` (see below).}

\item{Xcorr}{p by p matrix of correlations between variables
(covariates). When provided, it will be used to remove CSs whose
minimum correlation among variables is smaller than
\code{min_abs_corr}.}

\item{coverage}{A number between 0 and 1] specifying desired
coverage of each CS.}

\item{min_abs_corr}{a "purity" threshold for the CS. Any CS that
contains a pair of variables with correlation less than this
threshold will be filtered out and not reported.}

\item{dedup}{If \code{dedup = TRUE}, remove duplicate CSs.}

\item{squared}{If \code{squared = TRUE}, report min, mean and
median of squared correlation instead of the absolute correlation.}

\item{prune_by_cs}{Whether or not to ignore single effects not in
a reported CS when calculating PIP.}

\item{prior_tol}{Filter out effects having estimated prior variance
smaller than this threshold.}
}
\value{
\code{susie_get_objective} returns the evidence lower bound
(ELBO) achieved by the fitted susie model.
\code{susie_get_residual_variance} reeturns the (estimated or
fixed) residual variance parameter. \code{susie_get_prior_variance}
returns the (estimated or fixed) prior variance parameters.
\code{susie_get_posterior_mean} returns the posterior mean for the
regression coefficients of the fitted susie model.
\code{susie_get_posterior_sd} returns the posterior standard
deviation for coefficients of the fitted susie model.
\code{susie_get_niter} returns the number of model fitting
iterations performed. \code{susie_get_pip} returns a vector
containing the posterior inclusion probabilities (PIPs) for all
variables. \code{susie_get_lfsr} returns a vector containing the
average lfsr across variables for each single-effect, weighted by
the posterior inclusion probability (alpha).

A list with the following elements:

\item{cs}{a list, each element corresponds to a CS, and is a vector containing the indices of the variables in the CS.}

\item{coverage}{the nominal coverage specified for each CS.}

\item{purity}{(If `X` or `Xcorr` are provided), the purity of each CS.}

\item{cs_index}{(If `X` or `Xcorr` are provided) the index (in
  1,...,L) of each reported CS in the supplied susie fit.}
}
\description{
Give overview here.
}
\details{
Reports indices of variables in each credible set (CS)
  identified, as well as summaries of correlation among the variables
  included in each CS.  If desired, one can filter out CSs that do
  not meet a specified purity threshold (min_abs_corr); to do this
  either `X` or `Xcorr` must be supplied.
}
