% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_objective}
\alias{susie_get_objective}
\title{Get evidence lower bound (ELBO) from fitted susie model}
\usage{
susie_get_objective(res, last_only = TRUE, warning_tol = 1e-06)
}
\arguments{
\item{res}{A susie fit, e.g., the output of \code{\link{susie}}.}

\item{last_only}{whether or not to get ELBO from all iterations.}

\item{warning_tol}{warn if ELBO is non-decreasing by this tolerance
level.}
}
\description{
Get evidence lower bound (ELBO) from fitted susie model
}
