% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_get_PIP}
\alias{susie_get_PIP}
\title{Compute posterior inclusion probability (PIP) for all variables}
\usage{
susie_get_PIP(fitted, include_index = NULL)
}
\arguments{
\item{fitted}{a susie fit, the output of `susieR::susie()`, or simply the posterior
inclusion probability matrix alpha.}

\item{include_index}{index of single effect models to consider when calculating PIP. Default to NULL.}
}
\value{
a vector of posterior inclusion probability.
}
