% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{susie_init_coef}
\alias{susie_init_coef}
\title{Initialize a susie object using regression coefficients}
\usage{
susie_init_coef(coef_index, coef_value, p)
}
\arguments{
\item{coef_index}{An L-vector for indices of nonzero effects.}

\item{coef_value}{An L-vector for initial estimated beta values.}

\item{p}{A scalar the number of variables in the data.}
}
\value{
A list (alpha, mu, mu2) to be used by \code{susie}.
}
\description{
Initialize a susie object using regression coefficients
}
\examples{
# Add example(s) here.

}
