% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{susie_init_coef}
\alias{susie_init_coef}
\title{Initialize a susie object using regression coefficients}
\usage{
susie_init_coef(coef_index, coef_value, p)
}
\arguments{
\item{coef_index}{An L-vector containing the the indices of the
nonzero coefficients.}

\item{coef_value}{An L-vector containing initial coefficient
estimates.}

\item{p}{A scalar giving the number of variables.}
}
\value{
A list with elements \code{alpha}, \code{mu} and \code{mu2}
  to be used by \code{susie}.
}
\description{
Initialize a susie object using regression coefficients
}
\examples{
# Add example(s) here.

}
