% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_plots.R
\name{susie_plot}
\alias{susie_plot}
\alias{susie_plot_iteration}
\title{SuSiE Plots.}
\usage{
susie_plot(
  model,
  y,
  add_bar = FALSE,
  pos = NULL,
  b = NULL,
  max_cs = 400,
  add_legend = FALSE,
  ...
)

susie_plot_iteration(model, L, file_prefix, pos = NULL)
}
\arguments{
\item{model}{a susie fit, the output of `susieR::susie()`.
Multiple plots will be made for all iterations if `track_fit` was set to `TRUE` when running SuSiE.}

\item{y}{a string indicating what to plot: z (for z-score), PIP, log10PIP
or a random label to plot input data as is.}

\item{add_bar}{add horizontal bar to signals in credible interval.}

\item{pos}{index of variables to plot, default to all variables}

\item{b}{for simulated data, specify b = true effects (highlights in red).}

\item{max_cs}{the biggest CS to display, based on purity (set max_cs in between 0 and 1) or size (>1).}

\item{add_legend}{if TRUE, add a legend to annotate the size and purity of each CS discovered.}

\item{L}{an integer, number of CS to plot}

\item{file_prefix}{prefix to path of output plot file}
}
\description{
SuSiE Plots.
}
\details{
Plot per variable summary in SuSiE CSs.

Diagnostic plot for SuSiE iterations
}
