% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_plots.R
\name{susie_plot}
\alias{susie_plot}
\alias{susie_plot_iteration}
\title{SuSiE Plots.}
\usage{
susie_plot(
  model,
  y,
  add_bar = FALSE,
  pos = NULL,
  b = NULL,
  max_cs = 400,
  add_legend = FALSE,
  ...
)

susie_plot_iteration(model, L, file_prefix, pos = NULL)
}
\arguments{
\item{model}{A susie fit, typically an output from
\code{\link{susie}} or one of its variants. For \code{suse_plot},
the susie fit must have \code{model$z}, \code{model$PIP}, and may
include \code{model$sets}. \code{model} may also be a vector of
z-scores or PIPs.}

\item{y}{A string indicating what to plot: either \code{"z"} for
z-scores, \code{"PIP"} for posterior inclusion probabilities,
\code{"log10PIP"} for posterior inclusion probabiliities on the
(base-10) log-scale. For any other setting, the data are plotted as
is.}

\item{add_bar}{If \code{add_bar = TRUE}, add horizontal bar to
signals in credible interval.}

\item{pos}{Indices of variables to plot. If \code{pos = NULL} all
variables are plotted.}

\item{b}{For simulated data, set \code{b = TRUE} to highlight
"true" effects (highlights in red).}

\item{max_cs}{The largest credible set to display, either based on
purity (set \code{max_cs} between 0 and 1), or based on size (set
\code{max_cs > 1}).}

\item{add_legend}{If \code{add_legend = TRUE}, add a legend to
annotate the size and purity of each CS discovered.}

\item{\dots}{Additional arguments passed to
\code{\link[graphics]{plot}}.}

\item{L}{An integer specifying the number of credible sets to plot.}

\item{file_prefix}{Prefix to path of output plot file.}
}
\description{
\code{susie_plot} produces a per-variable summary of
  the SuSiE credible sets. \code{susie_plot_iteration} produces a
  diagnostic plot for the susie model fitting. For
  \code{susie_plot_iteration}, several plots will be created if
  \code{track_fit = TRUE} when calling \code{susie}.
}
\examples{
# Add example(s) here.

}
\seealso{
\code{\link{susie_plot_changepoint}}
}
