% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_utils.R
\name{susie_pplot}
\alias{susie_pplot}
\title{Plot per variable summary in SuSiE CSs}
\usage{
susie_pplot(data, res = NULL, dtype = "raw_data", CS = NULL,
  coverage = 0.9, add_bar = FALSE, pos = NULL, b = NULL, max_cs = 400,
  ...)
}
\arguments{
\item{data}{can be raw data \code{X} and \code{y} as \code{list(X=X,y=y)}, or be a vector of z-scores, or p-values.}

\item{res}{a susie fit, the output of `susieR::susie()`, or simply the posterior
inclusion probability matrix alpha.}

\item{dtype}{a string indicating the input data type (choices are raw_data (for raw data input), p (for p-value input), z (for z-scores input) and PIP (for PIP input))}

\item{CS}{the output of `susieR::susie_get_CS()`}

\item{coverage}{coverage of confident sets. Default to 0.9 for 90\% confidence interval.}

\item{add_bar}{add horizontal bar to signals in confidence interval.}

\item{b}{for simulated data, specify b = true effects (highlights in red).}

\item{max_cs}{the biggest CS to display, based on purity (set max_cs in between 0 and 1) or size (>1).}
}
\description{
Plot per variable summary in SuSiE CSs
}
