% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie_rss}
\alias{susie_rss}
\title{SuSiE with Regression Summary Statistics (RSS)}
\usage{
susie_rss(
  z = NULL,
  R,
  n = NULL,
  bhat = NULL,
  shat = NULL,
  var_y = NULL,
  L = min(10, ncol(R)),
  lambda = 0,
  maf = NULL,
  maf_thresh = 0,
  z_ld_weight = 0,
  prior_variance = 50,
  scaled_prior_variance = 0.2,
  residual_variance = NULL,
  prior_weights = NULL,
  null_weight = 0,
  standardize = TRUE,
  intercept_value = 0,
  estimate_residual_variance = FALSE,
  estimate_residual_method = c("MoM", "MLE"),
  estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM", "simple"),
  unmappable_effects = c("none", "inf"),
  check_null_threshold = 0,
  prior_tol = 1e-09,
  residual_variance_lowerbound = 0,
  residual_variance_upperbound = Inf,
  model_init = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 100,
  tol = 0.001,
  convergence_method = c("elbo", "pip"),
  verbose = FALSE,
  track_fit = FALSE,
  check_input = FALSE,
  check_prior = TRUE,
  check_R = TRUE,
  check_z = FALSE,
  n_purity = 100,
  r_tol = 1e-08,
  refine = FALSE
)
}
\arguments{
\item{z}{A p-vector of z-scores.}

\item{R}{A p by p correlation matrix.}

\item{n}{The sample size, not required but recommended.}

\item{bhat}{Alternative summary data giving the estimated effects
(a vector of length p). This, together with \code{shat}, may be
provided instead of \code{z}.}

\item{shat}{Alternative summary data giving the standard errors of
the estimated effects (a vector of length p). This, together with
\code{bhat}, may be provided instead of \code{z}.}

\item{lambda}{Regularization parameter for LD matrix. When
\code{lambda} > 0, you cannot use \code{unmappable_effects} methods.}

\item{z_ld_weight}{This parameter is included for backwards
compatibility with previous versions of the function, but it is no
longer recommended to set this to a non-zero value. When
\code{z_ld_weight > 0}, the matrix \code{R} is adjusted to be
\code{cov2cor((1-w)*R + w*tcrossprod(z))}, where \code{w =
z_ld_weight}.}

\item{estimate_residual_variance}{The default is FALSE, the
residual variance is fixed to 1 or variance of y. If the in-sample
LD matrix is provided, we recommend setting
\code{estimate_residual_variance = TRUE}.}

\item{check_R}{If TRUE, check that R is positive semidefinite.}

\item{check_z}{If TRUE, check that z lies in column space of R.}
}
\description{
Performs SuSiE regression using z-scores and correlation matrix.
Supports both standard RSS (lambda = 0) and RSS with regularized LD matrix (lambda > 0).
}
