% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_rss.R
\name{susie_rss}
\alias{susie_rss}
\title{Bayesian sum of single-effect (susie) linear regression using z scores}
\usage{
susie_rss(z, R, maf_thresh = 0, maf = NULL, L = 10, lambda = 0,
  prior_variance = 50, residual_variance = NULL, r_tol = 1e-08,
  prior_weights = NULL, null_weight = NULL, restrict = TRUE,
  estimate_residual_variance = TRUE, estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM"), max_iter = 100,
  s_init = NULL, intercept_value = 0, coverage = 0.95,
  min_abs_corr = 0.5, tol = 0.001, verbose = FALSE,
  track_fit = FALSE, check_input = FALSE)
}
\arguments{
\item{z}{a p vector of z scores.}

\item{R}{a p by p symmetric and positive semidefinite correlation matrix.}

\item{maf_thresh}{threshold for MAF}

\item{maf}{Minor Allele Frequency}

\item{L}{maximum number of non-zero effects}

\item{lambda}{fudge factor}

\item{prior_variance}{the prior variance (vector of length L, or scalar. In latter case gets repeated L times )}

\item{residual_variance}{the residual variance, a scaler between 0 and 1}

\item{r_tol}{tolerance level for eigen value check of positive semidefinite matrix of R.}

\item{prior_weights}{a p vector of prior probability that each element is non-zero}

\item{null_weight}{probability of no effect, for each single effect model}

\item{restrict}{whether to restrict the resiudal variance between 0 and 1}

\item{estimate_residual_variance}{indicates whether to estimate residual variance}

\item{estimate_prior_variance}{indicates whether to estimate prior}

\item{estimate_prior_method}{The method used for estimating prior variance, 'optim' or 'EM'}

\item{max_iter}{maximum number of iterations to perform}

\item{s_init}{a previous susie fit with which to initialize}

\item{intercept_value}{a value to assign to the intercept (since the intercept cannot be estimated from centered summary data). This
value will be used by coef.susie() to assign an intercept value, for consistency with the non-summary-statistic version of this function \code{susie}.
Set to NULL if you want coef.susie() not to include an intercept term (and so only return a p vector).}

\item{coverage}{coverage of confident sets. Default to 0.95 for 95\% credible interval.}

\item{min_abs_corr}{minimum of absolute value of correlation allowed in a credible set.}

\item{tol}{convergence tolerance based on alpha}

\item{verbose}{if true outputs some progress messages}

\item{track_fit}{add an attribute \code{trace} to output that saves current values of all iterations}

\item{check_input}{whether to skip the checks for R and z, the checks are:

1. Check whether R is positive semidefinite

2. Check whether z in space spanned by the non-zero eigenvectors of R}
}
\value{
a susie fit, which is a list with some or all of the following elements\cr
\item{alpha}{an L by p matrix of posterior inclusion probabilites}
\item{mu}{an L by p matrix of posterior means (conditional on inclusion)}
\item{mu2}{an L by p matrix of posterior second moments (conditional on inclusion)}
\item{Rr}{an p vector of t(X) times fitted values, the fitted values equal to X times colSums(alpha*mu))}
\item{sigma2}{residual variance}
\item{V}{prior variance}
}
\description{
Bayesian sum of single-effect (susie) linear regression using z scores
}
\details{
Performs sum of single-effect (susie) linear regression with z scores.
The summary data required are the p by p correlation matrix R, the p vector z. The summary stats should come from the same individuals.
This function fits the regression model z = sum_l Rb_l + e, where e is N(0,residual_variance * R) and the
sum_l b_l is a p vector of effects to be estimated.
The assumption is that each b_l has exactly one non-zero element, with all elements
equally likely to be non-zero. The prior on the non-zero element is N(0,var=prior_variance).
}
