% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie_set_init}
\alias{susie_set_init}
\title{initialize a susie object using given nonzero effect indices and beta values}
\usage{
susie_set_init(coef_index, coef_value, num_variables, V,
  residual_variance = NULL)
}
\arguments{
\item{coef_index}{a L-vector for indices of nonzero effects}

\item{coef_value}{a L-vector for initial estimated beta values}

\item{num_variables}{a scalar the number of variables in the data}

\item{V}{a scalar or an L-vector of prior variances}

\item{residual_variance}{a scalar containing residual variance}
}
\value{
a list of initialized alpha, mu, mu2, V and optionally sigma2
}
\description{
initialize a susie object using given nonzero effect indices and beta values
}
