% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie_ss}
\alias{susie_ss}
\title{SuSiE using Sufficient Statistics}
\usage{
susie_ss(
  XtX,
  Xty,
  yty,
  n,
  L = min(10, ncol(XtX)),
  X_colmeans = NA,
  y_mean = NA,
  maf = NULL,
  maf_thresh = 0,
  check_input = FALSE,
  r_tol = 1e-08,
  standardize = TRUE,
  scaled_prior_variance = 0.2,
  residual_variance = NULL,
  prior_weights = NULL,
  null_weight = 0,
  model_init = NULL,
  s_init = NULL,
  estimate_residual_variance = TRUE,
  estimate_residual_method = c("MoM", "MLE", "Servin_Stephens"),
  residual_variance_lowerbound = 0,
  residual_variance_upperbound = Inf,
  estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM", "simple"),
  unmappable_effects = c("none", "inf"),
  check_null_threshold = 0,
  prior_tol = 1e-09,
  max_iter = 100,
  tol = 0.001,
  convergence_method = c("elbo", "pip"),
  coverage = 0.95,
  min_abs_corr = 0.5,
  n_purity = 100,
  verbose = FALSE,
  track_fit = FALSE,
  check_prior = FALSE,
  refine = FALSE,
  alpha0 = 0.1,
  beta0 = 0.1
)
}
\arguments{
\item{XtX}{A p by p matrix, X'X, with columns of X centered to have mean zero.}

\item{Xty}{A p-vector, X'y, with y and columns of X centered to have mean zero.}

\item{yty}{A scalar, y'y, with y centered to have mean zero.}

\item{n}{The sample size.}

\item{L}{Maximum number of non-zero effects in the model. If L is larger than
the number of covariates, p, L is set to p.}

\item{X_colmeans}{A p-vector of column means of \code{X}. If both
\code{X_colmeans} and \code{y_mean} are provided, the intercept
is estimated; otherwise, the intercept is NA.}

\item{y_mean}{A scalar containing the mean of \code{y}. If both
\code{X_colmeans} and \code{y_mean} are provided, the intercept
is estimated; otherwise, the intercept is NA.}

\item{maf}{A p-vector of minor allele frequencies; to be used along with
\code{maf_thresh} to filter input summary statistics.}

\item{maf_thresh}{Variants with MAF smaller than this threshold are not used.}

\item{check_input}{If \code{check_input = TRUE}, \code{susie_ss} performs
additional checks on \code{XtX} and \code{Xty}. The checks are:
(1) check that \code{XtX} is positive semidefinite; (2) check that
\code{Xty} is in the space spanned by the non-zero eigenvectors of \code{XtX}.}

\item{r_tol}{Tolerance level for eigenvalue check of positive semidefinite
matrix \code{XtX}.}

\item{standardize}{If \code{standardize = TRUE}, standardize the
columns of X to unit variance prior to fitting (or equivalently
standardize XtX and Xty to have the same effect). Note that
\code{scaled_prior_variance} specifies the prior on the
coefficients of X \emph{after} standardization (if it is
performed). If you do not standardize, you may need to think more
carefully about specifying \code{scaled_prior_variance}. Whatever
your choice, the coefficients returned by \code{coef} are given for
\code{X} on the original input scale. Any column of \code{X} that
has zero variance is not standardized.}

\item{scaled_prior_variance}{The prior variance, divided by
\code{var(y)} (or by \code{(1/(n-1))yty} for
\code{susie_ss}); that is, the prior variance of each
non-zero element of b is \code{var(y) * scaled_prior_variance}. The
value provided should be either a scalar or a vector of length
\code{L}. If \code{estimate_prior_variance = TRUE}, this provides
initial estimates of the prior variances.}

\item{residual_variance}{Variance of the residual. If
\code{estimate_residual_variance = TRUE}, this value provides the
initial estimate of the residual variance. By default, it is set to
\code{var(y)} in \code{susie} and \code{(1/(n-1))yty} in
\code{susie_ss}.}

\item{prior_weights}{A vector of length p, in which each entry
gives the prior probability that corresponding column of X has a
nonzero effect on the outcome, y.}

\item{null_weight}{Prior probability of no effect (a number between 0 and 1,
and cannot be exactly 1).}

\item{model_init}{A previous susie fit with which to initialize.}

\item{s_init}{Deprecated; please use \code{model_init} instead. If
both \code{s_init} and \code{model_init} are provided, an error
is thrown.}

\item{estimate_residual_variance}{If
\code{estimate_residual_variance = TRUE}, the residual variance is
estimated, using \code{residual_variance} as an initial value. If
\code{estimate_residual_variance = FALSE}, the residual variance is
fixed to the value supplied by \code{residual_variance}.}

\item{estimate_residual_method}{The method used for estimating residual variance.
For the original SuSiE model, "MLE" and "MoM" estimation is equivalent, but for
the infinitesimal model, "MoM" is more stable. We recommend using "Servin_Stephens"
when n < 80 for improved coverage, although it is currently only implemented
for individual-level data.}

\item{residual_variance_lowerbound}{Lower limit on the estimated
residual variance. It is only relevant when
\code{estimate_residual_variance = TRUE}.}

\item{residual_variance_upperbound}{Upper limit on the estimated
residual variance. It is only relevant when
\code{estimate_residual_variance = TRUE}.}

\item{estimate_prior_variance}{If \code{estimate_prior_variance =
TRUE}, the prior variance is estimated (this is a separate
parameter for each of the L effects). If provided,
\code{scaled_prior_variance} is then used as an initial value for
the optimization. When \code{estimate_prior_variance = FALSE}, the
prior variance for each of the L effects is determined by the
value supplied to \code{scaled_prior_variance}.}

\item{estimate_prior_method}{The method used for estimating prior
variance. When \code{estimate_prior_method = "simple"} is used, the
likelihood at the specified prior variance is compared to the
likelihood at a variance of zero, and the setting with the larger
likelihood is retained.}

\item{unmappable_effects}{The method for modeling unmappable effects:
"none", "inf", "ash".}

\item{check_null_threshold}{When the prior variance is estimated,
compare the estimate with the null, and set the prior variance to
zero unless the log-likelihood using the estimate is larger by this
threshold amount. For example, if you set
\code{check_null_threshold = 0.1}, this will "nudge" the estimate
towards zero when the difference in log-likelihoods is small. A
note of caution that setting this to a value greater than zero may
lead the IBSS fitting procedure to occasionally decrease the ELBO. This
setting is disabled when using \code{unmappable_effects = "inf"} or
\code{unmappable_effects = "ash"}.}

\item{prior_tol}{When the prior variance is estimated, compare the
estimated value to \code{prior_tol} at the end of the computation,
and exclude a single effect from PIP computation if the estimated
prior variance is smaller than this tolerance value.}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{tol}{tol A small, non-negative number specifying the convergence
tolerance for the IBSS fitting procedure.}

\item{convergence_method}{When \code{converge_method = "elbo"} the fitting
procedure halts when the difference in the variational lower bound, or
\dQuote{ELBO} (the objective function to be maximized), is
less than \code{tol}. When \code{converge_method = "pip"} the fitting
procedure halts when the maximum absolute difference in \code{alpha} is less
than \code{tol}.}

\item{coverage}{A number between 0 and 1 specifying the
\dQuote{coverage} of the estimated confidence sets.}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
credible set. The default, 0.5, corresponds to a squared
correlation of 0.25, which is a commonly used threshold for
genotype data in genetic studies.}

\item{n_purity}{Passed as argument \code{n_purity} to
\code{\link{susie_get_cs}}.}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
a summary of the optimization settings, and refinement progress (if
\code{refine = TRUE}) are printed to the console.}

\item{track_fit}{If \code{track_fit = TRUE}, \code{trace}
is also returned containing detailed information about the
estimates at each iteration of the IBSS fitting procedure.}

\item{check_prior}{If \code{check_prior = TRUE}, it checks if the
estimated prior variance becomes unreasonably large (comparing with
10 * max(abs(z))^2).}

\item{refine}{If \code{refine = TRUE}, then an additional
iterative refinement procedure is used, after the IBSS algorithm,
to check and escape from local optima (see details).}

\item{alpha0}{Numerical parameter for the NIG prior when using
\code{estimate_residual_method = "Servin_Stephens"}.}

\item{beta0}{Numerical parameter for the NIG prior when using
\code{estimate_residual_method = "Servin_Stephens"}.}
}
\description{
Performs SuSiE regression using sufficient statistics (XtX, Xty,
yty, n) instead of individual-level data (X, y).
}
