% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss.R
\name{susie_ss}
\alias{susie_ss}
\title{Bayesian sum of single-effect (susie) linear regression using summary stat}
\usage{
susie_ss(XtX, Xty, var_y = 1, n, L = 10, scaled_prior_variance = 0.2,
  residual_variance = NULL, estimate_prior_variance = FALSE,
  max_iter = 100, s_init = NULL, verbose = FALSE, intercept = 0,
  tol = 1e-04)
}
\arguments{
\item{XtX}{a p by p matrix, X'X, where columns of X are centered to have mean 0}

\item{Xty}{a p vector, X'Y, where columns of X are centered and Y is centered to have mean 0}

\item{var_y}{the (sample) variance of the vector Y}

\item{n}{sample size}

\item{L}{maximum number of non-zero effects}

\item{scaled_prior_variance}{the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times )}

\item{residual_variance}{the residual variance (defaults to variance of Y)}

\item{estimate_prior_variance}{indicates whether to estimate prior (currently not recommended as not working as well)}

\item{max_iter}{maximum number of iterations to perform}

\item{s_init}{a previous susie fit with which to initialize}

\item{intercept}{a value to assign to the intercept (since the intercept cannot be estimated from centered summary data). This
value will be used by coef.susie() to assign an intercept value, for consistency with the non-summary-statistic version of this function \code{susie}.
Set to NULL if you want coef.susie() not to include an intercept term (and so only return a p vector).}

\item{tol}{convergence tolerance based on alpha}
}
\value{
a susie fit, which is a list with some or all of the following elements\cr
\item{alpha}{an L by p matrix of posterior inclusion probabilites}
\item{mu}{an L by p matrix of posterior means (conditional on inclusion)}
\item{mu2}{an L by p matrix of posterior second moments (conditional on inclusion)}
\item{XtXr}{an p vector of t(X) times fitted values, the fitted values equal to X times colSums(alpha*mu))}
\item{sigma2}{residual variance}
\item{V}{prior variance}
}
\description{
Bayesian sum of single-effect (susie) linear regression using summary stat
}
\details{
Performs sum of single-effect (susie) linear regression of Y on X when
only summary statistics are available. The summary data required are
the p by p matrix X'X, the p vector X'Y, and the sample variance of Y, or (1/n)Y'Y. The summary stats should come from the same individuals.
Both the columns of X and the vector Y
should be centered to have mean 0 before
computing these summary statistics; you may also want to scale each column of X and Y to have variance 1 (see examples).
This function fits the regression model Y = sum_l Xb_l + e, where elements of e are iid N(0,var=residual_variance) and the
sum_l b_l is a p vector of effects to be estimated.
The assumption is that each b_l has exactly one non-zero element, with all elements
equally likely to be non-zero. The prior on the non-zero element is N(0,var=scaled_prior_variance*var(Y)).
}
\examples{
set.seed(1)
n = 1000
p = 1000
beta = rep(0,p)
beta[1] = 1
beta[2] = 1
beta[3] = 1
beta[4] = 1
X = matrix(rnorm(n*p),nrow=n,ncol=p)
y = c(X \%*\% beta + rnorm(n))
X = scale(X,center=TRUE, scale=TRUE)
y = (y - mean(y))/sd(y)
res =susie_ss(XtX=t(X) \%*\% X,Xty= c(y \%*\% X), 1)
coef(res)
}
