% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie.R
\name{susie_ss}
\alias{susie_ss}
\title{SuSiE using Sufficient Statistics}
\usage{
susie_ss(
  XtX,
  Xty,
  yty,
  n,
  L = min(10, ncol(XtX)),
  X_colmeans = NA,
  y_mean = NA,
  maf = NULL,
  maf_thresh = 0,
  check_input = FALSE,
  r_tol = 1e-08,
  standardize = TRUE,
  scaled_prior_variance = 0.2,
  residual_variance = NULL,
  prior_weights = NULL,
  null_weight = 0,
  model_init = NULL,
  estimate_residual_variance = TRUE,
  estimate_residual_method = c("MLE", "MoM", "Servin_Stephens"),
  residual_variance_lowerbound = 0,
  residual_variance_upperbound = Inf,
  estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM", "simple"),
  unmappable_effects = c("none", "inf"),
  check_null_threshold = 0,
  prior_tol = 1e-09,
  max_iter = 100,
  tol = 0.001,
  convergence_method = c("elbo", "pip"),
  coverage = 0.95,
  min_abs_corr = 0.5,
  n_purity = 100,
  verbose = FALSE,
  track_fit = FALSE,
  check_prior = FALSE,
  refine = FALSE,
  ...
)
}
\arguments{
\item{XtX}{A p by p matrix, X'X, with columns of X centered to have mean zero.}

\item{Xty}{A p-vector, X'y, with y and columns of X centered to have mean zero.}

\item{yty}{A scalar, y'y, with y centered to have mean zero.}

\item{n}{The sample size.}

\item{X_colmeans}{A p-vector of column means of \code{X}. If both
\code{X_colmeans} and \code{y_mean} are provided, the intercept
is estimated; otherwise, the intercept is NA.}

\item{y_mean}{A scalar containing the mean of \code{y}. If both
\code{X_colmeans} and \code{y_mean} are provided, the intercept
is estimated; otherwise, the intercept is NA.}

\item{maf}{A p-vector of minor allele frequencies; to be used along with
\code{maf_thresh} to filter input summary statistics.}

\item{maf_thresh}{Variants with MAF smaller than this threshold are not used.}

\item{check_input}{If \code{check_input = TRUE}, \code{susie_ss} performs
additional checks on \code{XtX} and \code{Xty}. The checks are:
(1) check that \code{XtX} is positive semidefinite; (2) check that
\code{Xty} is in the space spanned by the non-zero eigenvectors of \code{XtX}.}

\item{r_tol}{Tolerance level for eigenvalue check of positive semidefinite
matrix \code{XtX}.}

\item{check_prior}{If \code{check_prior = TRUE}, it checks if the
estimated prior variance becomes unreasonably large (comparing with
10 * max(abs(z))^2).}
}
\description{
Performs SuSiE regression using sufficient statistics (XtX, Xty,
yty, n) instead of individual-level data (X, y).
}
