% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_trendfilter.R
\name{susie_trendfilter}
\alias{susie_trendfilter}
\title{perform trend filtering using SuSiE}
\usage{
susie_trendfilter(Y, order, ...)
}
\arguments{
\item{Y}{an n vector}

\item{order}{a scalar for the order of trend filtering}
}
\value{
SuSiE fit for trend filtering
}
\description{
perform trend filtering using SuSiE
}
