% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss_interface.R
\name{susie_z}
\alias{susie_z}
\title{Apply susie with z scores and correlation matrix R}
\usage{
susie_z(z, R, R_type = c("Cor", "Cov", "XtX"), n, var_y = 1, L = 10,
  scaled_prior_variance = 0.2, estimate_prior_variance = TRUE,
  prior_weights = NULL, null_weight = NULL, coverage = 0.95,
  min_abs_corr = 0.5, verbose = FALSE, track_fit = FALSE, ...)
}
\arguments{
\item{z}{a p vector of z scores.}

\item{R}{a p by p symmetric and positive semidefinite matrix}

\item{R_type}{the type of the matrix R, one of 'Cor', 'XtX' or 'Cov'}

\item{n}{sample size}

\item{var_y}{the (sample) variance of the y}

\item{L}{maximum number of non-zero effects}

\item{scaled_prior_variance}{the scaled prior variance (vector of length L, or scalar. In latter case gets repeated L times )}

\item{estimate_prior_variance}{indicates whether to estimate prior (currently not recommended as not working as well)}

\item{prior_weights}{a p vector of prior probability that each element is non-zero}

\item{null_weight}{probability of no effect, for each single effect model}

\item{coverage}{coverage of confident sets. Default to 0.95 for 95\% confidence interval.}

\item{min_abs_corr}{minimum of absolute value of correlation allowed in a confidence set.}

\item{verbose}{if true outputs some progress messages}

\item{track_fit}{add an attribute \code{trace} to output that saves current values of all iterations}

\item{...}{further arguments to be passed to \code{\link{susie_ss}}}
}
\value{
a susie fit
}
\description{
Apply susie with z scores and correlation matrix R
}
