% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_ss.R
\name{susie_z}
\alias{susie_z}
\title{Summary statistics version of SuSiE on z scores and correlation (or covariance) matrix.}
\usage{
susie_z(z, R, r_tol = 1e-08, L = 10,
  estimate_residual_variance = TRUE, optimV_method = c("optim", "EM",
  "uniroot"), prior_weights = NULL, null_weight = NULL,
  coverage = 0.95, min_abs_corr = 0.5, verbose = FALSE,
  track_fit = FALSE, ...)
}
\arguments{
\item{z}{a p vector of z scores.}

\item{R}{a p by p symmetric and positive semidefinite matrix. It can be X'X, covariance matrix or correlation matrix.}

\item{r_tol}{tolerance level for eigen value check of positive semidefinite matrix of R.}

\item{L}{maximum number of non-zero effects.}

\item{estimate_residual_variance}{indicates whether to estimate residual variance}

\item{optimV_method}{the method to estimate V, 'optim', 'EM' or 'uniroot'}

\item{prior_weights}{a p vector of prior probability that each element is non-zero.}

\item{null_weight}{probability of no effect, for each single effect model.}

\item{coverage}{coverage of confident sets. Default to 0.95 for 95\% credible interval.}

\item{min_abs_corr}{minimum of absolute value of correlation allowed in a credible set.}

\item{verbose}{if TRUE outputs some progress messages.}

\item{track_fit}{add an attribute \code{trace} to output that saves current values of all iterations.}

\item{...}{further arguments to be passed to \code{\link{susie_ss}}}
}
\value{
a susie fit
}
\description{
Summary statistics version of SuSiE on z scores and correlation (or covariance) matrix.
}
