% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_regression.R
\name{univariate_regression}
\alias{univariate_regression}
\title{Perform univariate regression between each column of X and y.
  Remove covariates if Z is not NULL.}
\usage{
univariate_regression(
  X,
  y,
  Z = NULL,
  center = TRUE,
  scale = FALSE,
  return_residuals = FALSE
)
}
\description{
Perform univariate regression between each column of X and y.
  Remove covariates if Z is not NULL.
}
