% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_vbupdate.R
\name{update_each_effect}
\alias{update_each_effect}
\title{update each effect once}
\usage{
update_each_effect(X, X.sparse, Y, cm, csd, s_init,
  estimate_prior_variance = FALSE)
}
\arguments{
\item{X}{an n by p matrix of covariantes, scaled}

\item{X.sparse}{an n by p sparse matrix of covariates, unscaled}

\item{Y}{an n vector of data}

\item{cm}{a p vector of column means}

\item{csd}{a p vector of column standard deviations}

\item{s_init}{a list with elements sigma2, sa2, alpha, mu, Xr}

\item{estimate_prior_variance}{says whether to estimate prior variance (sa2)}
}
\description{
update each effect once
}
