% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_each_effect.R
\name{update_each_effect}
\alias{update_each_effect}
\title{update each effect once}
\usage{
update_each_effect(X, Y, s, estimate_prior_variance = FALSE)
}
\arguments{
\item{X}{an n by p matrix of covariantes}

\item{Y}{an n vector of data}

\item{s}{a list with elements sigma2, V, alpha, mu, Xr}

\item{estimate_prior_variance}{boolean indicating whether to estimate prior variance}

\item{colSum}{of X^2}
}
\description{
update each effect once
}
