% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_each_effect_rss.R
\name{update_each_effect_rss}
\alias{update_each_effect_rss}
\title{update each effect once}
\usage{
update_each_effect_rss(
  R,
  z,
  s_init,
  Sigma,
  estimate_prior_variance = FALSE,
  estimate_prior_method = "optim",
  check_null_threshold = 0
)
}
\arguments{
\item{R}{a p by p symmetric and positive semidefinite correlation matrix.}

\item{z}{a p vector}

\item{s_init}{a list with elements sigma2, V, alpha, mu, Xr}

\item{Sigma}{sigma2*R + lambda I}

\item{estimate_prior_variance}{boolean indicating whether to estimate prior variance}

\item{check_null_threshold}{float a threshold on the log scale to compare likelihood between current estimate and zero the null}
}
\description{
update each effect once
}
\keyword{internal}
