% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_each_effect_ss.R
\name{update_each_effect_ss}
\alias{update_each_effect_ss}
\title{update each effect once}
\usage{
update_each_effect_ss(XtX, Xty, s_init, estimate_prior_variance = FALSE,
  optimV_method = c("EM", "optim", "uniroot"), niter)
}
\arguments{
\item{XtX}{a p by p matrix, t(X)X}

\item{Xty}{a p vector}

\item{s_init}{a list with elements sigma2, V, alpha, mu, Xr}

\item{estimate_prior_variance}{boolean indicating whether to estimate prior variance}

\item{optimV_method}{the method to estimate V, 'EM', 'optim', or 'uniroot'}
}
\description{
update each effect once
}
